@@ -1,6 +1,6 @@
 [tool.black]
 line-length = "110"
-target-version = ['py39']
+target-version = ['py310']
 include = '\.pyi?$'
 extend-exclude = '''
 (
@@ -21,7 +21,7 @@ extend-exclude = '''
 
 [tool.poetry]
 name = 'Modmail'
-version = '3.10.5'
+version = '4.1.1'
 description = "Modmail is similar to Reddit's Modmail, both in functionality and purpose. It serves as a shared inbox for server staff to communicate with their users in a seamless way."
 license = 'AGPL-3.0-only'
 authors = [
@@ -30,8 +30,8 @@ authors = [
     'Taki <noemail@example.com>'
 ]
 readme = 'README.md'
-repository = 'https://github.com/kyb3r/modmail'
-homepage = 'https://github.com/kyb3r/modmail'
+repository = 'https://github.com/modmail-dev/modmail'
+homepage = 'https://github.com/modmail-dev/modmail'
 keywords = ['discord', 'modmail']
 
 [tool.pylint.format]
