@@ -1,28 +1,28 @@
-__version__ = "3.10.5"
+__version__ = "4.1.1"
 
 
 import asyncio
 import copy
+import hashlib
 import logging
 import os
 import re
-import signal
 import string
+import struct
 import sys
+import platform
 import typing
-from datetime import datetime
+from datetime import datetime, timezone, timedelta
 from subprocess import PIPE
 from types import SimpleNamespace
 
 import discord
 import isodate
-from aiohttp import ClientSession
+from aiohttp import ClientSession, ClientResponseError
 from discord.ext import commands, tasks
 from discord.ext.commands.view import StringView
-from emoji import UNICODE_EMOJI
-from pkg_resources import parse_version
-
-from core.utils import tryint
+from emoji import is_emoji
+from packaging.version import Version
 
 
 try:
@@ -48,7 +48,7 @@ from core.models import (
 )
 from core.thread import ThreadManager
 from core.time import human_timedelta
-from core.utils import normalize_alias, truncate
+from core.utils import extract_block_timestamp, normalize_alias, parse_alias, truncate, tryint, human_join
 
 logger = getLogger(__name__)
 
@@ -65,32 +65,71 @@ if sys.platform == "win32":
 
 class ModmailBot(commands.Bot):
     def __init__(self):
-        intents = discord.Intents.all()
-        super().__init__(command_prefix=None, intents=intents)  # implemented in `get_prefix`
-        self._session = None
-        self._api = None
-        self.metadata_loop = None
-        self.autoupdate_loop = None
-        self.formatter = SafeFormatter()
-        self.loaded_cogs = ["cogs.modmail", "cogs.plugins", "cogs.utility"]
-        self._connected = asyncio.Event()
-        self.start_time = datetime.utcnow()
-        self._started = False
-
         self.config = ConfigManager(self)
         self.config.populate_cache()
 
+        intents = discord.Intents.all()
+        if not self.config["enable_presence_intent"]:
+            intents.presences = False
+
+        super().__init__(command_prefix=None, intents=intents)  # implemented in `get_prefix`
+        self.session = None
+        self._api = None
+        self.formatter = SafeFormatter()
+        self.loaded_cogs = ["cogs.modmail", "cogs.plugins", "cogs.utility"]
+        self._connected = None
+        self.start_time = discord.utils.utcnow()
+        self._started = False
+
         self.threads = ThreadManager(self)
 
-        self.log_file_name = os.path.join(temp_dir, f"{self.token.split('.')[0]}.log")
-        self._configure_logging()
+        log_dir = os.path.join(temp_dir, "logs")
+        if not os.path.exists(log_dir):
+            os.mkdir(log_dir)
+        self.log_file_path = os.path.join(log_dir, "modmail.log")
+        configure_logging(self)
 
         self.plugin_db = PluginDatabaseClient(self)  # Deprecated
         self.startup()
 
+    def get_guild_icon(
+        self, guild: typing.Optional[discord.Guild], *, size: typing.Optional[int] = None
+    ) -> str:
+        if guild is None:
+            guild = self.guild
+        if guild.icon is None:
+            return "https://cdn.discordapp.com/embed/avatars/0.png"
+        if size is None:
+            return guild.icon.url
+        return guild.icon.with_size(size).url
+
+    def _resolve_snippet(self, name: str) -> typing.Optional[str]:
+        """
+        Get actual snippet names from direct aliases to snippets.
+
+        If the provided name is a snippet, it's returned unchanged.
+        If there is an alias by this name, it is parsed to see if it
+        refers only to a snippet, in which case that snippet name is
+        returned.
+
+        If no snippets were found, None is returned.
+        """
+        if name in self.snippets:
+            return name
+
+        try:
+            (command,) = parse_alias(self.aliases[name])
+        except (KeyError, ValueError):
+            # There is either no alias by this name present or the
+            # alias has multiple steps.
+            pass
+        else:
+            if command in self.snippets:
+                return command
+
     @property
     def uptime(self) -> str:
-        now = datetime.utcnow()
+        now = discord.utils.utcnow()
         delta = now - self.start_time
         hours, remainder = divmod(int(delta.total_seconds()), 3600)
         minutes, seconds = divmod(remainder, 60)
@@ -133,47 +172,21 @@ class ModmailBot(commands.Bot):
         logger.info("discord.py: v%s", discord.__version__)
         logger.line()
 
+    async def load_extensions(self):
         for cog in self.loaded_cogs:
+            if cog in self.extensions:
+                continue
             logger.debug("Loading %s.", cog)
             try:
-                self.load_extension(cog)
+                await self.load_extension(cog)
                 logger.debug("Successfully loaded %s.", cog)
             except Exception:
                 logger.exception("Failed to load %s.", cog)
         logger.line("debug")
 
-    def _configure_logging(self):
-        level_text = self.config["log_level"].upper()
-        logging_levels = {
-            "CRITICAL": logging.CRITICAL,
-            "ERROR": logging.ERROR,
-            "WARNING": logging.WARNING,
-            "INFO": logging.INFO,
-            "DEBUG": logging.DEBUG,
-        }
-        logger.line()
-
-        log_level = logging_levels.get(level_text)
-        if log_level is None:
-            log_level = self.config.remove("log_level")
-            logger.warning("Invalid logging level set: %s.", level_text)
-            logger.warning("Using default logging level: INFO.")
-        else:
-            logger.info("Logging level: %s", level_text)
-
-        logger.info("Log file: %s", self.log_file_name)
-        configure_logging(self.log_file_name, log_level)
-        logger.debug("Successfully configured logging.")
-
     @property
     def version(self):
-        return parse_version(__version__)
-
-    @property
-    def session(self) -> ClientSession:
-        if self._session is None:
-            self._session = ClientSession(loop=self.loop)
-        return self._session
+        return Version(__version__)
 
     @property
     def api(self) -> ApiClient:
@@ -194,105 +207,73 @@ class ModmailBot(commands.Bot):
         return [self.prefix, f"<@{self.user.id}> ", f"<@!{self.user.id}> "]
 
     def run(self):
-        loop = self.loop
-
-        try:
-            loop.add_signal_handler(signal.SIGINT, lambda: loop.stop())
-            loop.add_signal_handler(signal.SIGTERM, lambda: loop.stop())
-        except NotImplementedError:
-            pass
-
         async def runner():
-            try:
-                retry_intents = False
+            async with self:
+                self._connected = asyncio.Event()
+                self.session = ClientSession(loop=self.loop)
+
+                if self.config["enable_presence_intent"]:
+                    logger.info("Starting bot with presence intent.")
+                else:
+                    logger.info("Starting bot without presence intent.")
+
                 try:
                     await self.start(self.token)
                 except discord.PrivilegedIntentsRequired:
-                    retry_intents = True
-                if retry_intents:
-                    await self.http.close()
-                    if self.ws is not None and self.ws.open:
-                        await self.ws.close(code=1000)
-                    self._ready.clear()
-                    intents = discord.Intents.default()
-                    intents.members = True
-                    # Try again with members intent
-                    self._connection._intents = intents
-                    logger.warning(
-                        "Attempting to login with only the server members privileged intent. Some plugins might not work correctly."
+                    logger.critical(
+                        "Privileged intents are not explicitly granted in the discord developers dashboard."
                     )
-                    await self.start(self.token)
-            except discord.PrivilegedIntentsRequired:
-                logger.critical(
-                    "Privileged intents are not explicitly granted in the discord developers dashboard."
-                )
-            except discord.LoginFailure:
-                logger.critical("Invalid token")
-            except Exception:
-                logger.critical("Fatal exception", exc_info=True)
-            finally:
-                if not self.is_closed():
-                    await self.close()
-                if self._session:
-                    await self._session.close()
+                except discord.LoginFailure:
+                    logger.critical("Invalid token")
+                except Exception:
+                    logger.critical("Fatal exception", exc_info=True)
+                finally:
+                    if self.session:
+                        await self.session.close()
+                    if not self.is_closed():
+                        await self.close()
 
-        # noinspection PyUnusedLocal
-        def stop_loop_on_completion(f):
-            loop.stop()
-
-        def _cancel_tasks():
-            if sys.version_info < (3, 8):
-                task_retriever = asyncio.Task.all_tasks
-            else:
+        async def _cancel_tasks():
+            async with self:
                 task_retriever = asyncio.all_tasks
+                loop = self.loop
+                tasks = {t for t in task_retriever() if not t.done() and t.get_coro() != cancel_tasks_coro}
 
-            tasks = {t for t in task_retriever(loop=loop) if not t.done()}
+                if not tasks:
+                    return
 
-            if not tasks:
-                return
+                logger.info("Cleaning up after %d tasks.", len(tasks))
+                for task in tasks:
+                    task.cancel()
 
-            logger.info("Cleaning up after %d tasks.", len(tasks))
-            for task in tasks:
-                task.cancel()
+                await asyncio.gather(*tasks, return_exceptions=True)
+                logger.info("All tasks finished cancelling.")
 
-            loop.run_until_complete(asyncio.gather(*tasks, return_exceptions=True))
-            logger.info("All tasks finished cancelling.")
+                for task in tasks:
+                    try:
+                        if task.exception() is not None:
+                            loop.call_exception_handler(
+                                {
+                                    "message": "Unhandled exception during Client.run shutdown.",
+                                    "exception": task.exception(),
+                                    "task": task,
+                                }
+                            )
+                    except (asyncio.InvalidStateError, asyncio.CancelledError):
+                        pass
 
-            for task in tasks:
-                if task.cancelled():
-                    continue
-                if task.exception() is not None:
-                    loop.call_exception_handler(
-                        {
-                            "message": "Unhandled exception during Client.run shutdown.",
-                            "exception": task.exception(),
-                            "task": task,
-                        }
-                    )
-
-        future = asyncio.ensure_future(runner(), loop=loop)
-        future.add_done_callback(stop_loop_on_completion)
         try:
-            loop.run_forever()
-        except KeyboardInterrupt:
+            asyncio.run(runner(), debug=bool(os.getenv("DEBUG_ASYNCIO")))
+        except (KeyboardInterrupt, SystemExit):
             logger.info("Received signal to terminate bot and event loop.")
         finally:
-            future.remove_done_callback(stop_loop_on_completion)
             logger.info("Cleaning up tasks.")
 
             try:
-                _cancel_tasks()
-                if sys.version_info >= (3, 6):
-                    loop.run_until_complete(loop.shutdown_asyncgens())
+                cancel_tasks_coro = _cancel_tasks()
+                asyncio.run(cancel_tasks_coro)
             finally:
                 logger.info("Closing the event loop.")
-
-        if not future.cancelled():
-            try:
-                return future.result()
-            except KeyboardInterrupt:
-                # I am unsure why this gets raised here but suppress it anyway
-                return None
 
     @property
     def bot_owner_ids(self):
@@ -523,6 +504,7 @@ class ModmailBot(commands.Bot):
         logger.debug("Connected to gateway.")
         await self.config.refresh()
         await self.api.setup_indexes()
+        await self.load_extensions()
         self._connected.set()
 
     async def on_ready(self):
@@ -557,6 +539,13 @@ class ModmailBot(commands.Bot):
             logger.info("Receiving guild ID: %s", self.modmail_guild.id)
         logger.line()
 
+        if "dev" in __version__:
+            logger.warning(
+                "You are running a developmental version. This should not be used in production. (v%s)",
+                __version__,
+            )
+            logger.line()
+
         await self.threads.populate_cache()
 
         # closures
@@ -565,7 +554,9 @@ class ModmailBot(commands.Bot):
         logger.line()
 
         for recipient_id, items in tuple(closures.items()):
-            after = (datetime.fromisoformat(items["time"]) - datetime.utcnow()).total_seconds()
+            after = (
+                datetime.fromisoformat(items["time"]).astimezone(timezone.utc) - discord.utils.utcnow()
+            ).total_seconds()
             if after <= 0:
                 logger.debug("Closing thread for recipient %s.", recipient_id)
                 after = 0
@@ -582,7 +573,7 @@ class ModmailBot(commands.Bot):
                 continue
 
             await thread.close(
-                closer=self.get_user(items["closer_id"]),
+                closer=await self.get_or_fetch_user(items["closer_id"]),
                 after=after,
                 silent=items["silent"],
                 delete_channel=items["delete_channel"],
@@ -598,13 +589,13 @@ class ModmailBot(commands.Bot):
                     {
                         "open": False,
                         "title": None,
-                        "closed_at": str(datetime.utcnow()),
+                        "closed_at": str(discord.utils.utcnow()),
                         "close_message": "Channel has been deleted, no closer found.",
                         "closer": {
                             "id": str(self.user.id),
                             "name": self.user.name,
                             "discriminator": self.user.discriminator,
-                            "avatar_url": str(self.user.avatar_url),
+                            "avatar_url": self.user.display_avatar.url,
                             "mod": True,
                         },
                     },
@@ -614,50 +605,42 @@ class ModmailBot(commands.Bot):
                 else:
                     logger.debug("Failed to close thread with channel %s, skipping.", log["channel_id"])
 
-        if self.config.get("data_collection"):
-            self.metadata_loop = tasks.Loop(
-                self.post_metadata,
-                seconds=0,
-                minutes=0,
-                hours=1,
-                count=None,
-                reconnect=True,
-                loop=None,
-            )
-            self.metadata_loop.before_loop(self.before_post_metadata)
-            self.metadata_loop.start()
-
-        self.autoupdate_loop = tasks.Loop(
-            self.autoupdate, seconds=0, minutes=0, hours=1, count=None, reconnect=True, loop=None
-        )
-        self.autoupdate_loop.before_loop(self.before_autoupdate)
-        self.autoupdate_loop.start()
-
         other_guilds = [guild for guild in self.guilds if guild not in {self.guild, self.modmail_guild}]
         if any(other_guilds):
             logger.warning(
                 "The bot is in more servers other than the main and staff server. "
                 "This may cause data compromise (%s).",
-                ", ".join(guild.name for guild in other_guilds),
+                ", ".join(str(guild.name) for guild in other_guilds),
             )
             logger.warning("If the external servers are valid, you may ignore this message.")
 
+        self.post_metadata.start()
+        self.autoupdate.start()
+        self.log_expiry.start()
         self._started = True
 
     async def convert_emoji(self, name: str) -> str:
         ctx = SimpleNamespace(bot=self, guild=self.modmail_guild)
         converter = commands.EmojiConverter()
 
-        if name not in UNICODE_EMOJI["en"]:
+        if not is_emoji(name):
             try:
                 name = await converter.convert(ctx, name.strip(":"))
             except commands.BadArgument as e:
-                logger.warning("%s is not a valid emoji. %s.", name, e)
+                logger.warning("%s is not a valid emoji: %s", name, e)
                 raise
         return name
 
-    async def retrieve_emoji(self) -> typing.Tuple[str, str]:
+    async def get_or_fetch_user(self, id: int) -> discord.User:
+        """
+        Retrieve a User based on their ID.
 
+        This tries getting the user from the cache and falls back to making
+        an API call if they're not found in the cache.
+        """
+        return self.get_user(id) or await self.fetch_user(id)
+
+    async def retrieve_emoji(self) -> typing.Tuple[str, str]:
         sent_emoji = self.config["sent_emoji"]
         blocked_emoji = self.config["blocked_emoji"]
 
@@ -681,7 +664,7 @@ class ModmailBot(commands.Bot):
 
     def check_account_age(self, author: discord.Member) -> bool:
         account_age = self.config.get("account_age")
-        now = datetime.utcnow()
+        now = discord.utils.utcnow()
 
         try:
             min_account_age = author.created_at + account_age
@@ -695,7 +678,7 @@ class ModmailBot(commands.Bot):
             logger.debug("Blocked due to account age, user %s.", author.name)
 
             if str(author.id) not in self.blocked_users:
-                new_reason = f"System Message: New Account. Required to wait for {delta}."
+                new_reason = f"System Message: New Account. User can try again {delta}."
                 self.blocked_users[str(author.id)] = new_reason
 
             return False
@@ -703,7 +686,7 @@ class ModmailBot(commands.Bot):
 
     def check_guild_age(self, author: discord.Member) -> bool:
         guild_age = self.config.get("guild_age")
-        now = datetime.utcnow()
+        now = discord.utils.utcnow()
 
         if not hasattr(author, "joined_at"):
             logger.warning("Not in guild, cannot verify guild_age, %s.", author.name)
@@ -721,7 +704,7 @@ class ModmailBot(commands.Bot):
             logger.debug("Blocked due to guild age, user %s.", author.name)
 
             if str(author.id) not in self.blocked_users:
-                new_reason = f"System Message: Recently Joined. Required to wait for {delta}."
+                new_reason = f"System Message: Recently Joined. User can try again {delta}."
                 self.blocked_users[str(author.id)] = new_reason
 
             return False
@@ -731,23 +714,14 @@ class ModmailBot(commands.Bot):
         if isinstance(author, discord.Member):
             for r in author.roles:
                 if str(r.id) in self.blocked_roles:
-
                     blocked_reason = self.blocked_roles.get(str(r.id)) or ""
-                    now = datetime.utcnow()
 
-                    # etc "blah blah blah... until 2019-10-14T21:12:45.559948."
-                    end_time = re.search(r"until ([^`]+?)\.$", blocked_reason)
-                    if end_time is None:
-                        # backwards compat
-                        end_time = re.search(r"%([^%]+?)%", blocked_reason)
-                        if end_time is not None:
-                            logger.warning(
-                                r"Deprecated time message for role %s, block and unblock again to update.",
-                                r.name,
-                            )
+                    try:
+                        end_time, after = extract_block_timestamp(blocked_reason, author.id)
+                    except ValueError:
+                        return False
 
                     if end_time is not None:
-                        after = (datetime.fromisoformat(end_time.group(1)) - now).total_seconds()
                         if after <= 0:
                             # No longer blocked
                             self.blocked_roles.pop(str(r.id))
@@ -763,26 +737,19 @@ class ModmailBot(commands.Bot):
             return True
 
         blocked_reason = self.blocked_users.get(str(author.id)) or ""
-        now = datetime.utcnow()
 
         if blocked_reason.startswith("System Message:"):
             # Met the limits already, otherwise it would've been caught by the previous checks
             logger.debug("No longer internally blocked, user %s.", author.name)
             self.blocked_users.pop(str(author.id))
             return True
-        # etc "blah blah blah... until 2019-10-14T21:12:45.559948."
-        end_time = re.search(r"until ([^`]+?)\.$", blocked_reason)
-        if end_time is None:
-            # backwards compat
-            end_time = re.search(r"%([^%]+?)%", blocked_reason)
-            if end_time is not None:
-                logger.warning(
-                    r"Deprecated time message for user %s, block and unblock again to update.",
-                    author.name,
-                )
+
+        try:
+            end_time, after = extract_block_timestamp(blocked_reason, author.id)
+        except ValueError:
+            return False
 
         if end_time is not None:
-            after = (datetime.fromisoformat(end_time.group(1)) - now).total_seconds()
             if after <= 0:
                 # No longer blocked
                 self.blocked_users.pop(str(author.id))
@@ -805,7 +772,6 @@ class ModmailBot(commands.Bot):
         channel: discord.TextChannel = None,
         send_message: bool = False,
     ) -> bool:
-
         member = self.guild.get_member(author.id)
         if member is None:
             # try to find in other guilds
@@ -852,7 +818,7 @@ class ModmailBot(commands.Bot):
 
     async def get_thread_cooldown(self, author: discord.Member):
         thread_cooldown = self.config.get("thread_cooldown")
-        now = datetime.utcnow()
+        now = discord.utils.utcnow()
 
         if thread_cooldown == isodate.Duration():
             return
@@ -870,10 +836,12 @@ class ModmailBot(commands.Bot):
             return
 
         try:
-            cooldown = datetime.fromisoformat(last_log_closed_at) + thread_cooldown
+            cooldown = datetime.fromisoformat(last_log_closed_at).astimezone(timezone.utc) + thread_cooldown
         except ValueError:
             logger.warning("Error with 'thread_cooldown'.", exc_info=True)
-            cooldown = datetime.fromisoformat(last_log_closed_at) + self.config.remove("thread_cooldown")
+            cooldown = datetime.fromisoformat(last_log_closed_at).astimezone(
+                timezone.utc
+            ) + self.config.remove("thread_cooldown")
 
         if cooldown > now:
             # User messaged before thread cooldown ended
@@ -889,7 +857,7 @@ class ModmailBot(commands.Bot):
         if reaction != "disable":
             try:
                 await msg.add_reaction(reaction)
-            except (discord.HTTPException, discord.InvalidArgument) as e:
+            except (discord.HTTPException, TypeError) as e:
                 logger.warning("Failed to add reaction %s: %s.", reaction, e)
                 return False
         return True
@@ -901,7 +869,7 @@ class ModmailBot(commands.Bot):
             return
         sent_emoji, blocked_emoji = await self.retrieve_emoji()
 
-        if message.type != discord.MessageType.default:
+        if message.type not in [discord.MessageType.default, discord.MessageType.reply]:
             return
 
         thread = await self.threads.find(recipient=message.author)
@@ -923,7 +891,10 @@ class ModmailBot(commands.Bot):
                     color=self.error_color,
                     description=self.config["disabled_new_thread_response"],
                 )
-                embed.set_footer(text=self.config["disabled_new_thread_footer"], icon_url=self.guild.icon_url)
+                embed.set_footer(
+                    text=self.config["disabled_new_thread_footer"],
+                    icon_url=self.get_guild_icon(guild=message.guild, size=128),
+                )
                 logger.info("A new thread was blocked from %s due to disabled Modmail.", message.author)
                 await self.add_reaction(message, blocked_emoji)
                 return await message.channel.send(embed=embed)
@@ -938,7 +909,7 @@ class ModmailBot(commands.Bot):
                 )
                 embed.set_footer(
                     text=self.config["disabled_current_thread_footer"],
-                    icon_url=self.guild.icon_url,
+                    icon_url=self.get_guild_icon(guild=message.guild, size=128),
                 )
                 logger.info("A message was blocked from %s due to disabled Modmail.", message.author)
                 await self.add_reaction(message, blocked_emoji)
@@ -963,6 +934,16 @@ class ModmailBot(commands.Bot):
                 await self.add_reaction(message, sent_emoji)
                 self.dispatch("thread_reply", thread, False, message, False, False)
 
+    def _get_snippet_command(self) -> commands.Command:
+        """Get the correct reply command based on the snippet config"""
+        modifiers = "f"
+        if self.config["plain_snippets"]:
+            modifiers += "p"
+        if self.config["anonymous_snippets"]:
+            modifiers += "a"
+
+        return self.get_command(f"{modifiers}reply")
+
     async def get_contexts(self, message, *, cls=commands.Context):
         """
         Returns all invocation contexts from the message.
@@ -973,7 +954,7 @@ class ModmailBot(commands.Bot):
         ctx = cls(prefix=self.prefix, view=view, bot=self, message=message)
         thread = await self.threads.find(channel=ctx.channel)
 
-        if self._skip_check(message.author.id, self.user.id):
+        if message.author.id == self.user.id:  # type: ignore
             return [ctx]
 
         prefixes = await self.get_prefix()
@@ -984,9 +965,18 @@ class ModmailBot(commands.Bot):
 
         invoker = view.get_word().lower()
 
+        # Check if a snippet is being called.
+        # This needs to be done before checking for aliases since
+        # snippets can have multiple words.
+        try:
+            # Use removeprefix once PY3.9+
+            snippet_text = self.snippets[message.content[len(invoked_prefix) :]]
+        except KeyError:
+            snippet_text = None
+
         # Check if there is any aliases being called.
         alias = self.aliases.get(invoker)
-        if alias is not None:
+        if alias is not None and snippet_text is None:
             ctxs = []
             aliases = normalize_alias(alias, message.content[len(f"{invoked_prefix}{invoker}") :])
             if not aliases:
@@ -994,18 +984,36 @@ class ModmailBot(commands.Bot):
                 self.aliases.pop(invoker)
 
             for alias in aliases:
-                view = StringView(invoked_prefix + alias)
+                command = None
+                try:
+                    snippet_text = self.snippets[alias]
+                except KeyError:
+                    command_invocation_text = alias
+                else:
+                    command = self._get_snippet_command()
+                    command_invocation_text = f"{invoked_prefix}{command} {snippet_text}"
+                view = StringView(invoked_prefix + command_invocation_text)
                 ctx_ = cls(prefix=self.prefix, view=view, bot=self, message=message)
                 ctx_.thread = thread
                 discord.utils.find(view.skip_string, prefixes)
                 ctx_.invoked_with = view.get_word().lower()
-                ctx_.command = self.all_commands.get(ctx_.invoked_with)
+                ctx_.command = command or self.all_commands.get(ctx_.invoked_with)
                 ctxs += [ctx_]
             return ctxs
 
         ctx.thread = thread
-        ctx.invoked_with = invoker
-        ctx.command = self.all_commands.get(invoker)
+
+        if snippet_text is not None:
+            # Process snippets
+            ctx.command = self._get_snippet_command()
+            reply_view = StringView(f"{invoked_prefix}{ctx.command} {snippet_text}")
+            discord.utils.find(reply_view.skip_string, prefixes)
+            ctx.invoked_with = reply_view.get_word().lower()
+            ctx.view = reply_view
+        else:
+            ctx.command = self.all_commands.get(invoker)
+            ctx.invoked_with = invoker
+
         return [ctx]
 
     async def trigger_auto_triggers(self, message, channel, *, cls=commands.Context):
@@ -1066,7 +1074,7 @@ class ModmailBot(commands.Bot):
         view = StringView(message.content)
         ctx = cls(prefix=self.prefix, view=view, bot=self, message=message)
 
-        if self._skip_check(message.author.id, self.user.id):
+        if message.author.id == self.user.id:
             return ctx
 
         ctx.thread = await self.threads.find(channel=ctx.channel)
@@ -1130,7 +1138,7 @@ class ModmailBot(commands.Bot):
                 color=self.main_color,
             )
             if self.config["show_timestamp"]:
-                em.timestamp = datetime.utcnow()
+                em.timestamp = discord.utils.utcnow()
 
             if not self.config["silent_alert_on_mention"]:
                 content = self.config["mention"]
@@ -1146,18 +1154,6 @@ class ModmailBot(commands.Bot):
 
         if isinstance(message.channel, discord.DMChannel):
             return await self.process_dm_modmail(message)
-
-        if message.content.startswith(self.prefix):
-            cmd = message.content[len(self.prefix) :].strip()
-
-            # Process snippets
-            cmd = cmd.lower()
-            if cmd in self.snippets:
-                snippet = self.snippets[cmd]
-                if self.config["anonymous_snippets"]:
-                    message.content = f"{self.prefix}fareply {snippet}"
-                else:
-                    message.content = f"{self.prefix}freply {snippet}"
 
         ctxs = await self.get_contexts(message)
         for ctx in ctxs:
@@ -1206,7 +1202,7 @@ class ModmailBot(commands.Bot):
             thread = await self.threads.find(recipient=user)
 
             if thread:
-                await thread.channel.trigger_typing()
+                await thread.channel.typing()
         else:
             if not self.config.get("mod_typing"):
                 return
@@ -1216,7 +1212,7 @@ class ModmailBot(commands.Bot):
                 for user in thread.recipients:
                     if await self.is_blocked(user):
                         continue
-                    await user.trigger_typing()
+                    await user.typing()
 
     async def handle_reaction_events(self, payload):
         user = self.get_user(payload.user_id)
@@ -1224,25 +1220,36 @@ class ModmailBot(commands.Bot):
             return
 
         channel = self.get_channel(payload.channel_id)
-        if not channel:  # dm channel not in internal cache
-            _thread = await self.threads.find(recipient=user)
-            if not _thread:
+        thread = None
+        # dm channel not in internal cache
+        if not channel:
+            thread = await self.threads.find(recipient=user)
+            if not thread:
                 return
-            channel = await _thread.recipient.create_dm()
+            channel = await thread.recipient.create_dm()
+            if channel.id != payload.channel_id:
+                return
 
+        from_dm = isinstance(channel, discord.DMChannel)
+        from_txt = isinstance(channel, discord.TextChannel)
+        if not from_dm and not from_txt:
+            return
+
+        if not thread:
+            params = {"recipient": user} if from_dm else {"channel": channel}
+            thread = await self.threads.find(**params)
+            if not thread:
+                return
+
+        # thread must exist before doing this API call
         try:
             message = await channel.fetch_message(payload.message_id)
         except (discord.NotFound, discord.Forbidden):
             return
 
         reaction = payload.emoji
-
         close_emoji = await self.convert_emoji(self.config["close_emoji"])
-
-        if isinstance(channel, discord.DMChannel):
-            thread = await self.threads.find(recipient=user)
-            if not thread:
-                return
+        if from_dm:
             if (
                 payload.event_type == "REACTION_ADD"
                 and message.embeds
@@ -1250,7 +1257,7 @@ class ModmailBot(commands.Bot):
                 and self.config.get("recipient_thread_close")
             ):
                 ts = message.embeds[0].timestamp
-                if thread and ts == thread.channel.created_at:
+                if ts == thread.channel.created_at:
                     # the reacted message is the corresponding thread creation embed
                     # closing thread
                     return await thread.close(closer=user)
@@ -1270,11 +1277,10 @@ class ModmailBot(commands.Bot):
                 logger.warning("Failed to find linked message for reactions: %s", e)
                 return
         else:
-            thread = await self.threads.find(channel=channel)
-            if not thread:
-                return
             try:
-                _, *linked_messages = await thread.find_linked_messages(message.id, either_direction=True)
+                _, *linked_messages = await thread.find_linked_messages(
+                    message1=message, either_direction=True
+                )
             except ValueError as e:
                 logger.warning("Failed to find linked message for reactions: %s", e)
                 return
@@ -1289,7 +1295,7 @@ class ModmailBot(commands.Bot):
                     for msg in linked_messages:
                         await msg.remove_reaction(reaction, self.user)
                     await message.remove_reaction(reaction, self.user)
-                except (discord.HTTPException, discord.InvalidArgument) as e:
+                except (discord.HTTPException, TypeError) as e:
                     logger.warning("Failed to remove reaction: %s", e)
 
     async def handle_react_to_contact(self, payload):
@@ -1320,7 +1326,7 @@ class ModmailBot(commands.Bot):
             )
             embed.set_footer(
                 text=self.config["disabled_new_thread_footer"],
-                icon_url=self.guild.icon_url,
+                icon_url=self.get_guild_icon(guild=channel.guild, size=128),
             )
             logger.info(
                 "A new thread using react to contact was blocked from %s due to disabled Modmail.",
@@ -1362,9 +1368,13 @@ class ModmailBot(commands.Bot):
             return
 
         audit_logs = self.modmail_guild.audit_logs(limit=10, action=discord.AuditLogAction.channel_delete)
-        entry = await audit_logs.find(lambda a: int(a.target.id) == channel.id)
+        found_entry = False
+        async for entry in audit_logs:
+            if int(entry.target.id) == channel.id:
+                found_entry = True
+                break
 
-        if entry is None:
+        if not found_entry:
             logger.debug("Cannot find the audit log entry for channel delete of %d.", channel.id)
             return
 
@@ -1378,28 +1388,44 @@ class ModmailBot(commands.Bot):
             await thread.close(closer=mod, silent=True, delete_channel=False)
 
     async def on_member_remove(self, member):
-        if member.guild != self.guild:
-            return
         thread = await self.threads.find(recipient=member)
         if thread:
-            if self.config["close_on_leave"]:
+            if member.guild == self.guild and self.config["close_on_leave"]:
                 await thread.close(
                     closer=member.guild.me,
                     message=self.config["close_on_leave_reason"],
                     silent=True,
                 )
             else:
-                embed = discord.Embed(
-                    description=self.config["close_on_leave_reason"], color=self.error_color
-                )
+                if len(self.guilds) > 1:
+                    guild_left = member.guild
+                    remaining_guilds = member.mutual_guilds
+
+                    if remaining_guilds:
+                        remaining_guild_names = [guild.name for guild in remaining_guilds]
+                        leave_message = (
+                            f"The recipient has left {guild_left}. "
+                            f"They are still in {human_join(remaining_guild_names, final='and')}."
+                        )
+                    else:
+                        leave_message = (
+                            f"The recipient has left {guild_left}. We no longer share any mutual servers."
+                        )
+                else:
+                    leave_message = "The recipient has left the server."
+
+                embed = discord.Embed(description=leave_message, color=self.error_color)
                 await thread.channel.send(embed=embed)
 
     async def on_member_join(self, member):
-        if member.guild != self.guild:
-            return
         thread = await self.threads.find(recipient=member)
         if thread:
-            embed = discord.Embed(description="The recipient has joined the server.", color=self.mod_color)
+            if len(self.guilds) > 1:
+                guild_joined = member.guild
+                join_message = f"The recipient has joined {guild_joined}."
+            else:
+                join_message = "The recipient has joined the server."
+            embed = discord.Embed(description=join_message, color=self.mod_color)
             await thread.channel.send(embed=embed)
 
     async def on_message_delete(self, message):
@@ -1422,7 +1448,13 @@ class ModmailBot(commands.Bot):
                 return
             message = message[0]
             embed = message.embeds[0]
-            embed.set_footer(text=f"{embed.footer.text} (deleted)", icon_url=embed.footer.icon_url)
+
+            if embed.footer.icon:
+                icon_url = embed.footer.icon.url
+            else:
+                icon_url = None
+
+            embed.set_footer(text=f"{embed.footer.text} (deleted)", icon_url=icon_url)
             await message.edit(embed=embed)
             return
 
@@ -1475,9 +1507,19 @@ class ModmailBot(commands.Bot):
         logger.error("Ignoring exception in %s.", event_method)
         logger.error("Unexpected exception:", exc_info=sys.exc_info())
 
-    async def on_command_error(self, context, exception):
+    async def on_command_error(
+        self, context: commands.Context, exception: Exception, *, unhandled_by_cog: bool = False
+    ) -> None:
+        if not unhandled_by_cog:
+            command = context.command
+            if command and command.has_error_handler():
+                return
+            cog = context.cog
+            if cog and cog.has_error_handler():
+                return
+
         if isinstance(exception, (commands.BadArgument, commands.BadUnionArgument)):
-            await context.trigger_typing()
+            await context.typing()
             await context.send(embed=discord.Embed(color=self.error_color, description=str(exception)))
         elif isinstance(exception, commands.CommandNotFound):
             logger.warning("CommandNotFound: %s", exception)
@@ -1512,21 +1554,23 @@ class ModmailBot(commands.Bot):
         else:
             logger.error("Unexpected exception:", exc_info=exception)
 
+    @tasks.loop(hours=1)
     async def post_metadata(self):
         info = await self.application_info()
 
+        delta = discord.utils.utcnow() - self.start_time
         data = {
             "bot_id": self.user.id,
             "bot_name": str(self.user),
-            "avatar_url": str(self.user.avatar_url),
+            "avatar_url": self.user.display_avatar.url,
             "guild_id": self.guild_id,
             "guild_name": self.guild.name,
             "member_count": len(self.guild.members),
-            "uptime": (datetime.utcnow() - self.start_time).total_seconds(),
+            "uptime": delta.total_seconds(),
             "latency": f"{self.ws.latency * 1000:.4f}",
             "version": str(self.version),
             "selfhosted": True,
-            "last_updated": str(datetime.utcnow()),
+            "last_updated": str(discord.utils.utcnow()),
         }
 
         if info.team is not None:
@@ -1543,56 +1587,71 @@ class ModmailBot(commands.Bot):
         async with self.session.post("https://api.modmail.dev/metadata", json=data):
             logger.debug("Uploading metadata to Modmail server.")
 
+    @post_metadata.before_loop
     async def before_post_metadata(self):
         await self.wait_for_connected()
+        if not self.config.get("data_collection") or not self.guild:
+            self.post_metadata.cancel()
+            return
+
         logger.debug("Starting metadata loop.")
         logger.line("debug")
-        if not self.guild:
-            self.metadata_loop.cancel()
 
+    @tasks.loop(hours=1)
     async def autoupdate(self):
         changelog = await Changelog.from_url(self)
         latest = changelog.latest_version
 
-        if self.version < parse_version(latest.version):
-            if self.hosting_method == HostingMethod.HEROKU:
+        if self.version < Version(latest.version):
+            error = None
+            data = {}
+            try:
+                # update fork if gh_token exists
                 data = await self.api.update_repository()
+            except InvalidConfigError:
+                pass
+            except ClientResponseError as exc:
+                error = exc
+            if self.hosting_method == HostingMethod.HEROKU:
+                if error is not None:
+                    logger.error(f"Autoupdate failed! Status: {error.status}.")
+                    logger.error(f"Error message: {error.message}")
+                    self.autoupdate.cancel()
+                    return
+
+                commit_data = data.get("data")
+                if not commit_data:
+                    return
+
+                logger.info("Bot has been updated.")
+
+                if not self.config["update_notifications"]:
+                    return
 
                 embed = discord.Embed(color=self.main_color)
-
-                commit_data = data["data"]
+                message = commit_data["commit"]["message"]
+                html_url = commit_data["html_url"]
+                short_sha = commit_data["sha"][:6]
                 user = data["user"]
+                embed.add_field(
+                    name="Merge Commit",
+                    value=f"[`{short_sha}`]({html_url}) " f"{message} - {user['username']}",
+                )
                 embed.set_author(
                     name=user["username"] + " - Updating Bot",
                     icon_url=user["avatar_url"],
                     url=user["url"],
                 )
 
-                embed.set_footer(text=f"Updating Modmail v{self.version} " f"-> v{latest.version}")
+                embed.set_footer(text=f"Updating Modmail v{self.version} -> v{latest.version}")
 
                 embed.description = latest.description
                 for name, value in latest.fields.items():
                     embed.add_field(name=name, value=value)
 
-                if commit_data:
-                    message = commit_data["commit"]["message"]
-                    html_url = commit_data["html_url"]
-                    short_sha = commit_data["sha"][:6]
-                    embed.add_field(
-                        name="Merge Commit",
-                        value=f"[`{short_sha}`]({html_url}) " f"{message} - {user['username']}",
-                    )
-                    logger.info("Bot has been updated.")
-                    channel = self.log_channel
-                    if self.config["update_notifications"]:
-                        await channel.send(embed=embed)
+                channel = self.update_channel
+                await channel.send(embed=embed)
             else:
-                try:
-                    # update fork if gh_token exists
-                    await self.api.update_repository()
-                except InvalidConfigError:
-                    pass
-
                 command = "git pull"
                 proc = await asyncio.create_subprocess_shell(
                     command,
@@ -1606,44 +1665,75 @@ class ModmailBot(commands.Bot):
 
                 if err and not res:
                     logger.warning(f"Autoupdate failed: {err}")
-                    self.autoupdate_loop.cancel()
+                    self.autoupdate.cancel()
                     return
 
                 elif res != "Already up to date.":
+                    if os.getenv("PIPENV_ACTIVE"):
+                        # Update pipenv if possible
+                        await asyncio.create_subprocess_shell(
+                            "pipenv sync",
+                            stderr=PIPE,
+                            stdout=PIPE,
+                        )
+                        message = ""
+                    else:
+                        message = "\n\nDo manually update dependencies if your bot has crashed."
+
                     logger.info("Bot has been updated.")
                     channel = self.update_channel
                     if self.hosting_method in (HostingMethod.PM2, HostingMethod.SYSTEMD):
                         embed = discord.Embed(title="Bot has been updated", color=self.main_color)
-                        embed.set_footer(text=f"Updating Modmail v{self.version} " f"-> v{latest.version}")
+                        embed.set_footer(
+                            text=f"Updating Modmail v{self.version} " f"-> v{latest.version} {message}"
+                        )
                         if self.config["update_notifications"]:
                             await channel.send(embed=embed)
                     else:
                         embed = discord.Embed(
                             title="Bot has been updated and is logging out.",
-                            description="If you do not have an auto-restart setup, please manually start the bot.",
+                            description=f"If you do not have an auto-restart setup, please manually start the bot. {message}",
                             color=self.main_color,
                         )
-                        embed.set_footer(text=f"Updating Modmail v{self.version} " f"-> v{latest.version}")
+                        embed.set_footer(text=f"Updating Modmail v{self.version} -> v{latest.version}")
                         if self.config["update_notifications"]:
                             await channel.send(embed=embed)
                     return await self.close()
 
+    @autoupdate.before_loop
     async def before_autoupdate(self):
         await self.wait_for_connected()
         logger.debug("Starting autoupdate loop")
 
         if self.config.get("disable_autoupdates"):
             logger.warning("Autoupdates disabled.")
-            self.autoupdate_loop.cancel()
+            self.autoupdate.cancel()
+            return
 
         if self.hosting_method == HostingMethod.DOCKER:
             logger.warning("Autoupdates disabled as using Docker.")
-            self.autoupdate_loop.cancel()
+            self.autoupdate.cancel()
+            return
 
         if not self.config.get("github_token") and self.hosting_method == HostingMethod.HEROKU:
             logger.warning("GitHub access token not found.")
             logger.warning("Autoupdates disabled.")
-            self.autoupdate_loop.cancel()
+            self.autoupdate.cancel()
+            return
+
+    @tasks.loop(hours=1, reconnect=False)
+    async def log_expiry(self):
+        log_expire_after = self.config.get("log_expiration")
+        if log_expire_after == isodate.Duration():
+            return self.log_expiry.stop()
+
+        now = discord.utils.utcnow()
+        expiration_datetime = now - log_expire_after
+        # WARNING: comparison is done lexicographically, not by date.
+        # This is fine as long as the date is in zero-padded ISO format, which it should be.
+        expired_logs = await self.db.logs.delete_many({"closed_at": {"$lte": str(expiration_datetime)}})
+
+        logger.info(f"Deleted {expired_logs.deleted_count} expired logs.")
 
     def format_channel_name(self, author, exclude_channel=None, force_null=False):
         """Sanitises a username for use with text channel names
@@ -1654,18 +1744,28 @@ class ModmailBot(commands.Bot):
         if force_null:
             name = new_name = "null"
         else:
-            if self.config["use_user_id_channel_name"]:
+            if self.config["use_random_channel_name"]:
+                to_hash = self.token.split(".")[-1] + str(author.id)
+                digest = hashlib.md5(to_hash.encode("utf8"), usedforsecurity=False)
+                name = new_name = digest.hexdigest()[-8:]
+            elif self.config["use_user_id_channel_name"]:
                 name = new_name = str(author.id)
             elif self.config["use_timestamp_channel_name"]:
                 name = new_name = author.created_at.isoformat(sep="-", timespec="minutes")
             else:
-                name = author.name.lower()
+                if self.config["use_nickname_channel_name"]:
+                    author_member = self.guild.get_member(author.id)
+                    name = author_member.display_name.lower()
+                else:
+                    name = author.name.lower()
+
                 if force_null:
                     name = "null"
 
-                name = new_name = (
-                    "".join(l for l in name if l not in string.punctuation and l.isprintable()) or "null"
-                ) + f"-{author.discriminator}"
+                name = "".join(l for l in name if l not in string.punctuation and l.isprintable()) or "null"
+                if author.discriminator != "0":
+                    name += f"-{author.discriminator}"
+                new_name = name
 
         counter = 1
         existed = set(c.name for c in guild.text_channels if c != exclude_channel)
@@ -1679,17 +1779,42 @@ class ModmailBot(commands.Bot):
 def main():
     try:
         # noinspection PyUnresolvedReferences
-        import uvloop
+        import uvloop  # type: ignore
 
         logger.debug("Setting up with uvloop.")
         uvloop.install()
     except ImportError:
         pass
 
+    try:
+        import cairosvg  # noqa: F401
+    except OSError:
+        if os.name == "nt":
+            if struct.calcsize("P") * 8 != 64:
+                logger.error(
+                    "Unable to import cairosvg, ensure your Python is a 64-bit version: https://www.python.org/downloads/"
+                )
+            else:
+                logger.error(
+                    "Unable to import cairosvg, install GTK Installer for Windows and restart your system (https://github.com/tschoonj/GTK-for-Windows-Runtime-Environment-Installer/releases/latest)"
+                )
+        else:
+            if "ubuntu" in platform.version().lower() or "debian" in platform.version().lower():
+                logger.error(
+                    "Unable to import cairosvg, try running `sudo apt-get install libpangocairo-1.0-0` or report on our support server with your OS details: https://discord.gg/etJNHCQ"
+                )
+            else:
+                logger.error(
+                    "Unable to import cairosvg, report on our support server with your OS details: https://discord.gg/etJNHCQ"
+                )
+        sys.exit(0)
+
     # check discord version
-    if discord.__version__ != "1.7.3":
+    discord_version = "2.3.2"
+    if discord.__version__ != discord_version:
         logger.error(
-            "Dependencies are not updated, run pipenv install. discord.py version expected 1.7.3, received %s",
+            "Dependencies are not updated, run pipenv install. discord.py version expected %s, received %s",
+            discord_version,
             discord.__version__,
         )
         sys.exit(0)
