@@ -6,8 +6,8 @@ import shutil
 import sys
 import typing
 import zipfile
-from importlib import invalidate_caches
 from difflib import get_close_matches
+from importlib import invalidate_caches
 from pathlib import Path, PurePath
 from re import match
 from site import USER_SITE
@@ -15,13 +15,12 @@ from subprocess import PIPE
 
 import discord
 from discord.ext import commands
-
-from pkg_resources import parse_version
+from packaging.version import Version
 
 from core import checks
 from core.models import PermissionLevel, getLogger
 from core.paginator import EmbedPaginatorSession
-from core.utils import truncate, trigger_typing
+from core.utils import trigger_typing, truncate
 
 logger = getLogger(__name__)
 
@@ -114,7 +113,7 @@ class Plugins(commands.Cog):
     These addons could have a range of features from moderation to simply
     making your life as a moderator easier!
     Learn how to create a plugin yourself here:
-    https://github.com/kyb3r/modmail/wiki/Plugins
+    https://github.com/modmail-dev/modmail/wiki/Plugins
     """
 
     def __init__(self, bot):
@@ -123,21 +122,22 @@ class Plugins(commands.Cog):
         self.loaded_plugins = set()
         self._ready_event = asyncio.Event()
 
-        self.bot.loop.create_task(self.populate_registry())
-
+    async def cog_load(self):
+        await self.populate_registry()
         if self.bot.config.get("enable_plugins"):
-            self.bot.loop.create_task(self.initial_load_plugins())
+            await self.initial_load_plugins()
         else:
             logger.info("Plugins not loaded since ENABLE_PLUGINS=false.")
 
     async def populate_registry(self):
-        url = "https://raw.githubusercontent.com/kyb3r/modmail/be14cca89742cd691f6b91f0d0931a074ad81e3b/plugins/registry.json"
-        async with self.bot.session.get(url) as resp:
-            self.registry = json.loads(await resp.text())
+        url = "https://raw.githubusercontent.com/modmail-dev/modmail/master/plugins/registry.json"
+        try:
+            async with self.bot.session.get(url) as resp:
+                self.registry = json.loads(await resp.text())
+        except asyncio.TimeoutError:
+            logger.warning("Failed to fetch registry. Loading with empty registry")
 
     async def initial_load_plugins(self):
-        await self.bot.wait_for_connected()
-
         for plugin_name in list(self.bot.config["plugins"]):
             try:
                 plugin = Plugin.from_string(plugin_name, strict=True)
@@ -258,7 +258,7 @@ class Plugins(commands.Cog):
                 sys.path.insert(0, USER_SITE)
 
         try:
-            self.bot.load_extension(plugin.ext_string)
+            await self.bot.load_extension(plugin.ext_string)
             logger.info("Loaded plugin: %s", plugin.ext_string.split(".")[-1])
             self.loaded_plugins.add(plugin)
 
@@ -266,8 +266,18 @@ class Plugins(commands.Cog):
             logger.error("Plugin load failure: %s", plugin.ext_string, exc_info=True)
             raise InvalidPluginError("Cannot load extension, plugin invalid.") from exc
 
-    async def parse_user_input(self, ctx, plugin_name, check_version=False):
+    async def unload_plugin(self, plugin: Plugin) -> None:
+        try:
+            await self.bot.unload_extension(plugin.ext_string)
+        except commands.ExtensionError as exc:
+            raise exc
 
+        ext_parent = ".".join(plugin.ext_string.split(".")[:-1])
+        for module in list(sys.modules.keys()):
+            if module == ext_parent or module.startswith(ext_parent + "."):
+                del sys.modules[module]
+
+    async def parse_user_input(self, ctx, plugin_name, check_version=False):
         if not self.bot.config["enable_plugins"]:
             embed = discord.Embed(
                 description="Plugins are disabled, enable them by setting `ENABLE_PLUGINS=true`",
@@ -292,7 +302,7 @@ class Plugins(commands.Cog):
             if check_version:
                 required_version = details.get("bot_version", False)
 
-                if required_version and self.bot.version < parse_version(required_version):
+                if required_version and self.bot.version < Version(required_version):
                     embed = discord.Embed(
                         description="Your bot's version is too low. "
                         f"This plugin requires version `{required_version}`.",
@@ -304,6 +314,14 @@ class Plugins(commands.Cog):
             plugin = Plugin(user, repo, plugin_name, branch)
 
         else:
+            if self.bot.config.get("registry_plugins_only"):
+                embed = discord.Embed(
+                    description="This plugin is not in the registry. To install this plugin, "
+                    "you must set `REGISTRY_PLUGINS_ONLY=no` or remove this key in your .env file.",
+                    color=self.bot.error_color,
+                )
+                await ctx.send(embed=embed)
+                return
             try:
                 plugin = Plugin.from_string(plugin_name)
             except InvalidPluginError:
@@ -372,7 +390,7 @@ class Plugins(commands.Cog):
             logger.warning("Unable to download plugin %s.", plugin, exc_info=True)
 
             embed = discord.Embed(
-                description=f"Failed to download plugin, check logs for error.\n{type(e)}: {e}",
+                description=f"Failed to download plugin, check logs for error.\n{type(e).__name__}: {e}",
                 color=self.bot.error_color,
             )
 
@@ -382,7 +400,6 @@ class Plugins(commands.Cog):
         await self.bot.config.update()
 
         if self.bot.config.get("enable_plugins"):
-
             invalidate_caches()
 
             try:
@@ -391,7 +408,7 @@ class Plugins(commands.Cog):
                 logger.warning("Unable to load plugin %s.", plugin, exc_info=True)
 
                 embed = discord.Embed(
-                    description=f"Failed to download plugin, check logs for error.\n{type(e)}: {e}",
+                    description=f"Failed to load plugin, check logs for error.\n{type(e).__name__}: {e}",
                     color=self.bot.error_color,
                 )
 
@@ -432,7 +449,7 @@ class Plugins(commands.Cog):
 
         if self.bot.config.get("enable_plugins"):
             try:
-                self.bot.unload_extension(plugin.ext_string)
+                await self.unload_plugin(plugin)
                 self.loaded_plugins.remove(plugin)
             except (commands.ExtensionNotLoaded, KeyError):
                 logger.warning("Plugin was never loaded.")
@@ -474,9 +491,10 @@ class Plugins(commands.Cog):
             await self.download_plugin(plugin, force=True)
             if self.bot.config.get("enable_plugins"):
                 try:
-                    self.bot.unload_extension(plugin.ext_string)
+                    await self.unload_plugin(plugin)
                 except commands.ExtensionError:
                     logger.warning("Plugin unload fail.", exc_info=True)
+
                 try:
                     await self.load_plugin(plugin)
                 except Exception:
@@ -484,12 +502,12 @@ class Plugins(commands.Cog):
                         description=f"Failed to update {plugin.name}. This plugin will now be removed from your bot.",
                         color=self.bot.error_color,
                     )
-                    self.bot.config["plugins"].remove(plugin_name)
-                    logger.debug("Failed to update %s. Removed plugin from config.", plugin_name)
+                    self.bot.config["plugins"].remove(str(plugin))
+                    logger.debug("Failed to update %s. Removed plugin from config.", plugin)
                 else:
-                    logger.debug("Updated %s.", plugin_name)
+                    logger.debug("Updated %s.", plugin)
             else:
-                logger.debug("Updated %s.", plugin_name)
+                logger.debug("Updated %s.", plugin)
             return await ctx.send(embed=embed)
 
     @plugins.command(name="update")
@@ -523,17 +541,20 @@ class Plugins(commands.Cog):
         for ext in list(self.bot.extensions):
             if not ext.startswith("plugins."):
                 continue
+            logger.error("Unloading plugin: %s.", ext)
             try:
-                logger.error("Unloading plugin: %s.", ext)
-                self.bot.unload_extension(ext)
-            except Exception:
-                logger.error("Failed to unload plugin: %s.", ext)
-            else:
-                if not self.loaded_plugins:
-                    continue
                 plugin = next((p for p in self.loaded_plugins if p.ext_string == ext), None)
                 if plugin:
+                    await self.unload_plugin(plugin)
                     self.loaded_plugins.remove(plugin)
+                else:
+                    await self.bot.unload_extension(ext)
+            except Exception:
+                logger.error("Failed to unload plugin: %s.", ext)
+
+        for module in list(sys.modules.keys()):
+            if module.startswith("plugins."):
+                del sys.modules[module]
 
         self.bot.config["plugins"].clear()
         await self.bot.config.update()
@@ -619,6 +640,14 @@ class Plugins(commands.Cog):
 
         registry = sorted(self.registry.items(), key=lambda elem: elem[0])
 
+        if not registry:
+            embed = discord.Embed(
+                color=self.bot.error_color,
+                description="Registry is empty. This could be because it failed to load.",
+            )
+            await ctx.send(embed=embed)
+            return
+
         if isinstance(plugin_name, int):
             index = plugin_name - 1
             if index < 0:
@@ -669,7 +698,7 @@ class Plugins(commands.Cog):
                 embed.set_footer(text="This plugin is currently loaded.")
             else:
                 required_version = details.get("bot_version", False)
-                if required_version and self.bot.version < parse_version(required_version):
+                if required_version and self.bot.version < Version(required_version):
                     embed.set_footer(
                         text="Your bot is unable to install this plugin, "
                         f"minimum required version is v{required_version}."
@@ -730,12 +759,12 @@ class Plugins(commands.Cog):
 
         for page in pages:
             embed = discord.Embed(color=self.bot.main_color, description=page)
-            embed.set_author(name="Plugin Registry", icon_url=self.bot.user.avatar_url)
+            embed.set_author(name="Plugin Registry", icon_url=self.bot.user.display_avatar.url)
             embeds.append(embed)
 
         paginator = EmbedPaginatorSession(ctx, *embeds)
         await paginator.run()
 
 
-def setup(bot):
-    bot.add_cog(Plugins(bot))
+async def setup(bot):
+    await bot.add_cog(Plugins(bot))
