@@ -5,22 +5,21 @@ import random
 import re
 import traceback
 from contextlib import redirect_stdout
-from datetime import datetime
 from difflib import get_close_matches
 from io import BytesIO, StringIO
 from itertools import takewhile, zip_longest
 from json import JSONDecodeError, loads
 from subprocess import PIPE
 from textwrap import indent
-from types import SimpleNamespace
 from typing import Union
 
 import discord
-from aiohttp import ClientResponseError
 from discord.enums import ActivityType, Status
 from discord.ext import commands, tasks
 from discord.ext.commands.view import StringView
-from pkg_resources import parse_version
+
+from aiohttp import ClientResponseError
+from packaging.version import Version
 
 from core import checks, utils
 from core.changelog import Changelog
@@ -31,7 +30,7 @@ from core.models import (
     UnseenFormatter,
     getLogger,
 )
-from core.utils import trigger_typing, truncate
+from core.utils import trigger_typing, truncate, DummyParam
 from core.paginator import EmbedPaginatorSession, MessagePaginatorSession
 
 
@@ -40,7 +39,7 @@ logger = getLogger(__name__)
 
 class ModmailHelpCommand(commands.HelpCommand):
     async def command_callback(self, ctx, *, command=None):
-        """Ovrwrites original command_callback to ensure `help` without any arguments
+        """Overwrites original command_callback to ensure `help` without any arguments
         returns with checks, `help all` returns without checks"""
         if command is None:
             self.verify_checks = True
@@ -54,7 +53,7 @@ class ModmailHelpCommand(commands.HelpCommand):
 
     async def format_cog_help(self, cog, *, no_cog=False):
         bot = self.context.bot
-        prefix = self.clean_prefix
+        prefix = self.context.clean_prefix
 
         formats = [""]
         for cmd in await self.filter_commands(
@@ -90,19 +89,23 @@ class ModmailHelpCommand(commands.HelpCommand):
 
             embed.add_field(name="Commands", value=format_ or "No commands.")
 
-            continued = " (Continued)" if embeds else ""
             name = cog.qualified_name + " - Help" if not no_cog else "Miscellaneous Commands"
-            embed.set_author(name=name + continued, icon_url=bot.user.avatar_url)
+            embed.set_author(name=name, icon_url=bot.user.display_avatar.url)
 
             embed.set_footer(
                 text=f'Type "{prefix}{self.command_attrs["name"]} command" '
                 "for more info on a specific command."
             )
             embeds.append(embed)
+
+        if len(embeds) > 1:
+            for n, em in enumerate(embeds):
+                em.set_author(name=f"{em.author.name} [{n + 1}]", icon_url=em.author.icon_url)
+
         return embeds
 
     def process_help_msg(self, help_: str):
-        return help_.format(prefix=self.clean_prefix) if help_ else "No help message."
+        return help_.format(prefix=self.context.clean_prefix) if help_ else "No help message."
 
     async def send_bot_help(self, mapping):
         embeds = []
@@ -139,7 +142,7 @@ class ModmailHelpCommand(commands.HelpCommand):
             perm_level = "NONE"
 
         embed = discord.Embed(
-            title=f"`{self.get_command_signature(topic)}`",
+            title=f"`{self.get_command_signature(topic).strip()}`",
             color=self.context.bot.main_color,
             description=self.process_help_msg(topic.help),
         )
@@ -174,7 +177,7 @@ class ModmailHelpCommand(commands.HelpCommand):
 
         embed.add_field(name="Sub Command(s)", value=format_[:1024], inline=False)
         embed.set_footer(
-            text=f'Type "{self.clean_prefix}{self.command_attrs["name"]} command" '
+            text=f'Type "{self.context.clean_prefix}{self.command_attrs["name"]} command" '
             "for more info on a command."
         )
 
@@ -185,7 +188,18 @@ class ModmailHelpCommand(commands.HelpCommand):
         val = self.context.bot.snippets.get(command)
         if val is not None:
             embed = discord.Embed(title=f"{command} is a snippet.", color=self.context.bot.main_color)
-            embed.add_field(name=f"`{command}` will send:", value=val)
+            embed.add_field(name=f"`{command}` will send:", value=val, inline=False)
+
+            snippet_aliases = []
+            for alias in self.context.bot.aliases:
+                if self.context.bot._resolve_snippet(alias) == command:
+                    snippet_aliases.append(f"`{alias}`")
+
+            if snippet_aliases:
+                embed.add_field(
+                    name="Aliases to this snippet:", value=",".join(snippet_aliases), inline=False
+                )
+
             return await self.get_destination().send(embed=embed)
 
         val = self.context.bot.aliases.get(command)
@@ -216,7 +230,7 @@ class ModmailHelpCommand(commands.HelpCommand):
                         embed.add_field(name=f"Step {i}:", value=val)
 
             embed.set_footer(
-                text=f'Type "{self.clean_prefix}{self.command_attrs["name"]} alias" '
+                text=f'Type "{self.context.clean_prefix}{self.command_attrs["name"]} alias" '
                 "for more details on aliases."
             )
             return await self.get_destination().send(embed=embed)
@@ -238,7 +252,7 @@ class ModmailHelpCommand(commands.HelpCommand):
         else:
             embed.title = "Cannot find command or category"
             embed.set_footer(
-                text=f'Type "{self.clean_prefix}{self.command_attrs["name"]}" '
+                text=f'Type "{self.context.clean_prefix}{self.command_attrs["name"]}" '
                 "for a list of all available commands."
             )
         await self.get_destination().send(embed=embed)
@@ -257,10 +271,12 @@ class Utility(commands.Cog):
             },
         )
         self.bot.help_command.cog = self
-        self.loop_presence.start()  # pylint: disable=no-member
         if not self.bot.config.get("enable_eval"):
             self.eval_.enabled = False
             logger.info("Eval disabled. enable_eval=False")
+
+    async def cog_load(self):
+        self.loop_presence.start()  # pylint: disable=no-member
 
     def cog_unload(self):
         self.bot.help_command = self._original_help_command
@@ -303,13 +319,13 @@ class Utility(commands.Cog):
     @utils.trigger_typing
     async def about(self, ctx):
         """Shows information about this bot."""
-        embed = discord.Embed(color=self.bot.main_color, timestamp=datetime.utcnow())
+        embed = discord.Embed(color=self.bot.main_color, timestamp=discord.utils.utcnow())
         embed.set_author(
             name="Modmail - About",
-            icon_url=self.bot.user.avatar_url,
+            icon_url=self.bot.user.display_avatar.url,
             url="https://discord.gg/F34cRU8",
         )
-        embed.set_thumbnail(url=self.bot.user.avatar_url)
+        embed.set_thumbnail(url=self.bot.user.display_avatar.url)
 
         desc = "This is an open source Discord bot that serves as a means for "
         desc += "members to easily communicate with server administrators in "
@@ -326,17 +342,17 @@ class Utility(commands.Cog):
         latest = changelog.latest_version
 
         if self.bot.version.is_prerelease:
-            stable = next(filter(lambda v: not parse_version(v.version).is_prerelease, changelog.versions))
+            stable = next(filter(lambda v: not Version(v.version).is_prerelease, changelog.versions))
             footer = f"You are on the prerelease version • the latest version is v{stable.version}."
-        elif self.bot.version < parse_version(latest.version):
+        elif self.bot.version < Version(latest.version):
             footer = f"A newer version is available v{latest.version}."
         else:
             footer = "You are up to date with the latest version."
 
         embed.add_field(
             name="Want Modmail in Your Server?",
-            value="Follow the installation guide on [GitHub](https://github.com/kyb3r/modmail/) "
-            "and join our [Discord server](https://discord.gg/F34cRU8)!",
+            value="Follow the installation guide on [GitHub](https://github.com/modmail-dev/modmail/) "
+            "and join our [Discord server](https://discord.gg/cnUpwrnpYb)!",
             inline=False,
         )
 
@@ -364,7 +380,7 @@ class Utility(commands.Cog):
         """Shows the sponsors of this project."""
 
         async with self.bot.session.get(
-            "https://raw.githubusercontent.com/kyb3r/modmail/master/SPONSORS.json"
+            "https://raw.githubusercontent.com/modmail-dev/modmail/master/SPONSORS.json"
         ) as resp:
             data = loads(await resp.text())
 
@@ -385,13 +401,7 @@ class Utility(commands.Cog):
     async def debug(self, ctx):
         """Shows the recent application logs of the bot."""
 
-        log_file_name = self.bot.token.split(".")[0]
-
-        with open(
-            os.path.join(os.path.dirname(os.path.abspath(__file__)), f"../temp/{log_file_name}.log"),
-            "r+",
-            encoding="utf-8",
-        ) as f:
+        with open(self.bot.log_file_path, "r+", encoding="utf-8") as f:
             logs = f.read().strip()
 
         if not logs:
@@ -400,7 +410,7 @@ class Utility(commands.Cog):
                 title="Debug Logs:",
                 description="You don't have any logs at the moment.",
             )
-            embed.set_footer(text="Go to Heroku to see your logs.")
+            embed.set_footer(text="Go to your console to see your logs.")
             return await ctx.send(embed=embed)
 
         messages = []
@@ -417,7 +427,7 @@ class Utility(commands.Cog):
                     msg = "```Haskell\n"
             msg += line
             if len(msg) + 3 > 2000:
-                msg = msg[:1993] + "[...]```"
+                msg = msg[:1992] + "[...]```"
                 messages.append(msg)
                 msg = "```Haskell\n"
 
@@ -439,12 +449,8 @@ class Utility(commands.Cog):
         """Posts application-logs to Hastebin."""
 
         haste_url = os.environ.get("HASTE_URL", "https://hastebin.cc")
-        log_file_name = self.bot.token.split(".")[0]
 
-        with open(
-            os.path.join(os.path.dirname(os.path.abspath(__file__)), f"../temp/{log_file_name}.log"),
-            "rb+",
-        ) as f:
+        with open(self.bot.log_file_path, "rb+") as f:
             logs = BytesIO(f.read().strip())
 
         try:
@@ -466,7 +472,7 @@ class Utility(commands.Cog):
                 color=self.bot.main_color,
                 description="Something's wrong. We're unable to upload your logs to hastebin.",
             )
-            embed.set_footer(text="Go to Heroku to see your logs.")
+            embed.set_footer(text="Go to your console to see your logs.")
         await ctx.send(embed=embed)
 
     @debug.command(name="clear", aliases=["wipe"])
@@ -475,12 +481,7 @@ class Utility(commands.Cog):
     async def debug_clear(self, ctx):
         """Clears the locally cached logs."""
 
-        log_file_name = self.bot.token.split(".")[0]
-
-        with open(
-            os.path.join(os.path.dirname(os.path.abspath(__file__)), f"../temp/{log_file_name}.log"),
-            "w",
-        ):
+        with open(self.bot.log_file_path, "w"):
             pass
         await ctx.send(
             embed=discord.Embed(color=self.bot.main_color, description="Cached logs are now cleared.")
@@ -498,6 +499,7 @@ class Utility(commands.Cog):
             - `listening`
             - `watching`
             - `competing`
+            - `custom`
 
         When activity type is set to `listening`,
         it must be followed by a "to": "listening to..."
@@ -508,6 +510,9 @@ class Utility(commands.Cog):
         When activity type is set to `streaming`, you can set
         the linked twitch page:
         - `{prefix}config set twitch_url https://www.twitch.tv/somechannel/`
+
+        When activity type is set to `custom`, you can set
+        any custom text as the activity message.
 
         To remove the current activity status:
         - `{prefix}activity clear`
@@ -521,12 +526,12 @@ class Utility(commands.Cog):
             return await ctx.send(embed=embed)
 
         if not message:
-            raise commands.MissingRequiredArgument(SimpleNamespace(name="message"))
+            raise commands.MissingRequiredArgument(DummyParam("message"))
 
         try:
             activity_type = ActivityType[activity_type]
         except KeyError:
-            raise commands.MissingRequiredArgument(SimpleNamespace(name="activity"))
+            raise commands.MissingRequiredArgument(DummyParam("activity"))
 
         activity, _ = await self.set_presence(activity_type=activity_type, activity_message=message)
 
@@ -571,7 +576,7 @@ class Utility(commands.Cog):
         try:
             status = Status[status_type]
         except KeyError:
-            raise commands.MissingRequiredArgument(SimpleNamespace(name="status"))
+            raise commands.MissingRequiredArgument(DummyParam("status"))
 
         _, status = await self.set_presence(status=status)
 
@@ -583,7 +588,6 @@ class Utility(commands.Cog):
         return await ctx.send(embed=embed)
 
     async def set_presence(self, *, status=None, activity_type=None, activity_message=None):
-
         if status is None:
             status = self.bot.config.get("status")
 
@@ -609,7 +613,9 @@ class Utility(commands.Cog):
         elif activity_type == ActivityType.streaming:
             url = self.bot.config["twitch_url"]
 
-        if activity_type is not None:
+        if activity_type == ActivityType.custom:
+            activity = discord.CustomActivity(name=activity_message)
+        elif activity_type is not None:
             activity = discord.Activity(type=activity_type, name=activity_message, url=url)
         else:
             activity = None
@@ -794,7 +800,7 @@ class Utility(commands.Cog):
 
         if key in keys:
             try:
-                self.bot.config.set(key, value)
+                await self.bot.config.set(key, value)
                 await self.bot.config.update()
                 embed = discord.Embed(
                     title="Success",
@@ -848,7 +854,7 @@ class Utility(commands.Cog):
             if key in keys:
                 desc = f"`{key}` is set to `{self.bot.config[key]}`"
                 embed = discord.Embed(color=self.bot.main_color, description=desc)
-                embed.set_author(name="Config variable", icon_url=self.bot.user.avatar_url)
+                embed.set_author(name="Config variable", icon_url=self.bot.user.display_avatar.url)
 
             else:
                 embed = discord.Embed(
@@ -865,7 +871,7 @@ class Utility(commands.Cog):
                 color=self.bot.main_color,
                 description="Here is a list of currently set configuration variable(s).",
             )
-            embed.set_author(name="Current config(s):", icon_url=self.bot.user.avatar_url)
+            embed.set_author(name="Current config(s):", icon_url=self.bot.user.display_avatar.url)
             config = self.bot.config.filter_default(self.bot.config)
 
             for name, value in config.items():
@@ -913,9 +919,7 @@ class Utility(commands.Cog):
         for i, (current_key, info) in enumerate(config_help.items()):
             if current_key == key:
                 index = i
-            embed = discord.Embed(
-                title=f"Configuration description on {current_key}:", color=self.bot.main_color
-            )
+            embed = discord.Embed(title=f"{current_key}", color=self.bot.main_color)
             embed.add_field(name="Default:", value=fmt(info["default"]), inline=False)
             embed.add_field(name="Information:", value=fmt(info["description"]), inline=False)
             if info["examples"]:
@@ -1006,7 +1010,7 @@ class Utility(commands.Cog):
                 color=self.bot.error_color, description="You dont have any aliases at the moment."
             )
             embed.set_footer(text=f'Do "{self.bot.prefix}help alias" for more commands.')
-            embed.set_author(name="Aliases", icon_url=ctx.guild.icon_url)
+            embed.set_author(name="Aliases", icon_url=self.bot.get_guild_icon(guild=ctx.guild, size=128))
             return await ctx.send(embed=embed)
 
         embeds = []
@@ -1014,7 +1018,9 @@ class Utility(commands.Cog):
         for i, names in enumerate(zip_longest(*(iter(sorted(self.bot.aliases)),) * 15)):
             description = utils.format_description(i, names)
             embed = discord.Embed(color=self.bot.main_color, description=description)
-            embed.set_author(name="Command Aliases", icon_url=ctx.guild.icon_url)
+            embed.set_author(
+                name="Command Aliases", icon_url=self.bot.get_guild_icon(guild=ctx.guild, size=128)
+            )
             embeds.append(embed)
 
         session = EmbedPaginatorSession(ctx, *embeds)
@@ -1071,7 +1077,9 @@ class Utility(commands.Cog):
             linked_command = view.get_word().lower()
             message = view.read_rest()
 
-            if not self.bot.get_command(linked_command):
+            is_snippet = val in self.bot.snippets
+
+            if not self.bot.get_command(linked_command) and not is_snippet:
                 alias_command = self.bot.aliases.get(linked_command)
                 if alias_command is not None:
                     save_aliases.extend(utils.normalize_alias(alias_command, message))
@@ -1099,7 +1107,7 @@ class Utility(commands.Cog):
         await self.bot.config.update()
         return embed
 
-    @alias.command(name="add")
+    @alias.command(name="add", aliases=["create", "make"])
     @checks.has_permissions(PermissionLevel.MODERATOR)
     async def alias_add(self, ctx, name: str.lower, *, value):
         """
@@ -1595,7 +1603,10 @@ class Utility(commands.Cog):
                                 for name, level in takewhile(lambda x: x is not None, items)
                             )
                             embed = discord.Embed(color=self.bot.main_color, description=description)
-                            embed.set_author(name="Permission Overrides", icon_url=ctx.guild.icon_url)
+                            embed.set_author(
+                                name="Permission Overrides",
+                                icon_url=self.bot.get_guild_icon(guild=ctx.guild, size=128),
+                            )
                             embeds.append(embed)
 
                     session = EmbedPaginatorSession(ctx, *embeds)
@@ -1915,7 +1926,7 @@ class Utility(commands.Cog):
     async def update(self, ctx, *, flag: str = ""):
         """
         Update Modmail.
-        To stay up-to-date with the latest commit rom GitHub, specify "force" as the flag.
+        To stay up-to-date with the latest commit from GitHub, specify "force" as the flag.
         """
 
         changelog = await Changelog.from_url(self.bot)
@@ -1923,10 +1934,10 @@ class Utility(commands.Cog):
 
         desc = (
             f"The latest version is [`{self.bot.version}`]"
-            "(https://github.com/kyb3r/modmail/blob/master/bot.py#L25)"
+            "(https://github.com/modmail-dev/modmail/blob/master/bot.py#L1)"
         )
 
-        if self.bot.version >= parse_version(latest.version) and flag.lower() != "force":
+        if self.bot.version >= Version(latest.version) and flag.lower() != "force":
             embed = discord.Embed(title="Already up to date", description=desc, color=self.bot.main_color)
 
             data = await self.bot.api.get_user_info()
@@ -1935,16 +1946,39 @@ class Utility(commands.Cog):
                 embed.set_author(name=user["username"], icon_url=user["avatar_url"], url=user["url"])
             await ctx.send(embed=embed)
         else:
-            if self.bot.hosting_method == HostingMethod.HEROKU:
+            error = None
+            data = {}
+            try:
+                # update fork if gh_token exists
                 data = await self.bot.api.update_repository()
+            except InvalidConfigError:
+                pass
+            except ClientResponseError as exc:
+                error = exc
+
+            if self.bot.hosting_method == HostingMethod.HEROKU:
+                if error is not None:
+                    embed = discord.Embed(
+                        title="Update failed",
+                        description=f"Error status: {error.status}.\nError message: {error.message}",
+                        color=self.bot.error_color,
+                    )
+                    return await ctx.send(embed=embed)
+                if not data:
+                    # invalid gh_token
+                    embed = discord.Embed(
+                        title="Update failed",
+                        description="Invalid Github token.",
+                        color=self.bot.error_color,
+                    )
+                    return await ctx.send(embed=embed)
 
                 commit_data = data["data"]
                 user = data["user"]
-
                 if commit_data and commit_data.get("html_url"):
                     embed = discord.Embed(color=self.bot.main_color)
 
-                    embed.set_footer(text=f"Updating Modmail v{self.bot.version} " f"-> v{latest.version}")
+                    embed.set_footer(text=f"Updating Modmail v{self.bot.version} -> v{latest.version}")
 
                     embed.set_author(
                         name=user["username"] + " - Updating bot",
@@ -1962,21 +1996,14 @@ class Utility(commands.Cog):
                 else:
                     embed = discord.Embed(
                         title="Already up to date",
-                        description="No further updates required",
+                        description="No further updates required.",
                         color=self.bot.main_color,
                     )
                     embed.set_footer(text="Force update")
                     embed.set_author(name=user["username"], icon_url=user["avatar_url"], url=user["url"])
                 await ctx.send(embed=embed)
             else:
-                # update fork if gh_token exists
-                try:
-                    await self.bot.api.update_repository()
-                except InvalidConfigError:
-                    pass
-
                 command = "git pull"
-
                 proc = await asyncio.create_subprocess_shell(
                     command,
                     stderr=PIPE,
@@ -2100,5 +2127,5 @@ class Utility(commands.Cog):
         await self.bot.add_reaction(ctx.message, "\u2705")
 
 
-def setup(bot):
-    bot.add_cog(Utility(bot))
+async def setup(bot):
+    await bot.add_cog(Utility(bot))
