@@ -1,9 +1,9 @@
 import secrets
 import sys
-from datetime import datetime
 from json import JSONDecodeError
-from typing import Union, Optional
+from typing import Any, Dict, Union, Optional
 
+import discord
 from discord import Member, DMChannel, TextChannel, Message
 from discord.ext import commands
 
@@ -19,6 +19,7 @@ logger = getLogger(__name__)
 class GitHub:
     """
     The client for interacting with GitHub API.
+
     Parameters
     ----------
     bot : Bot
@@ -31,6 +32,7 @@ class GitHub:
         URL to the avatar in GitHub.
     url : str, optional
         URL to the GitHub profile.
+
     Attributes
     ----------
     bot : Bot
@@ -43,6 +45,7 @@ class GitHub:
         URL to the avatar in GitHub.
     url : str
         URL to the GitHub profile.
+
     Class Attributes
     ----------------
     BASE : str
@@ -60,10 +63,10 @@ class GitHub:
     """
 
     BASE = "https://api.github.com"
-    REPO = BASE + "/repos/kyb3r/modmail"
+    REPO = BASE + "/repos/modmail-dev/modmail"
     MERGE_URL = BASE + "/repos/{username}/modmail/merges"
     FORK_URL = REPO + "/forks"
-    STAR_URL = BASE + "/user/starred/kyb3r/modmail"
+    STAR_URL = BASE + "/user/starred/modmail-dev/modmail"
 
     def __init__(self, bot, access_token: str = "", username: str = "", **kwargs):
         self.bot = bot
@@ -77,7 +80,7 @@ class GitHub:
             self.headers = {"Authorization": "token " + str(access_token)}
 
     @property
-    def BRANCH(self):
+    def BRANCH(self) -> str:
         return "master" if not self.bot.version.is_prerelease else "development"
 
     async def request(
@@ -85,11 +88,13 @@ class GitHub:
         url: str,
         method: str = "GET",
         payload: dict = None,
-        return_response: bool = False,
         headers: dict = None,
-    ) -> Union[ClientResponse, dict, str]:
+        return_response: bool = False,
+        read_before_return: bool = False,
+    ) -> Union[ClientResponse, Dict[str, Any], str]:
         """
         Makes a HTTP request.
+
         Parameters
         ----------
         url : str
@@ -98,16 +103,20 @@ class GitHub:
             The HTTP method (POST, GET, PUT, DELETE, FETCH, etc.).
         payload : Dict[str, Any]
             The json payload to be sent along the request.
-        return_response : bool
-            Whether the `ClientResponse` object should be returned.
         headers : Dict[str, str]
             Additional headers to `headers`.
+        return_response : bool
+            Whether the `ClientResponse` object should be returned.
+        read_before_return : bool
+            Whether to perform `.read()` method before returning the `ClientResponse` object.
+            Only valid if `return_response` is set to `True`.
+
         Returns
         -------
         ClientResponse or Dict[str, Any] or List[Any] or str
             `ClientResponse` if `return_response` is `True`.
-            `dict` if the returned data is a json object.
-            `list` if the returned data is a json list.
+            `Dict[str, Any]` if the returned data is a json object.
+            `List[Any]` if the returned data is a json list.
             `str` if the returned data is not a valid json data,
             the raw response.
         """
@@ -117,19 +126,32 @@ class GitHub:
             headers = self.headers
         async with self.session.request(method, url, headers=headers, json=payload) as resp:
             if return_response:
+                if read_before_return:
+                    await resp.read()
                 return resp
-            try:
-                return await resp.json()
-            except (JSONDecodeError, ClientResponseError):
-                return await resp.text()
 
-    def filter_valid(self, data):
+            return await self._get_response_data(resp)
+
+    @staticmethod
+    async def _get_response_data(response: ClientResponse) -> Union[Dict[str, Any], str]:
+        """
+        Internal method to convert the response data to `dict` if the data is a
+        json object, or to `str` (raw response) if the data is not a valid json.
+        """
+        try:
+            return await response.json()
+        except (JSONDecodeError, ClientResponseError):
+            return await response.text()
+
+    def filter_valid(self, data) -> Dict[str, Any]:
         """
         Filters configuration keys that are accepted.
+
         Parameters
         ----------
         data : Dict[str, Any]
             The data that needs to be cleaned.
+
         Returns
         -------
         Dict[str, Any]
@@ -138,42 +160,79 @@ class GitHub:
         valid_keys = self.bot.config.valid_keys.difference(self.bot.config.protected_keys)
         return {k: v for k, v in data.items() if k in valid_keys}
 
-    async def update_repository(self, sha: str = None) -> Optional[dict]:
+    async def update_repository(self, sha: str = None) -> Dict[str, Any]:
         """
         Update the repository from Modmail main repo.
+
         Parameters
         ----------
-        sha : Optional[str], optional
-            The commit SHA to update the repository.
+        sha : Optional[str]
+            The commit SHA to update the repository. If `None`, the latest
+            commit SHA will be fetched.
+
         Returns
         -------
-        Optional[dict]
-            If the response is a dict.
+        Dict[str, Any]
+            A dictionary that contains response data.
         """
         if not self.username:
             raise commands.CommandInvokeError("Username not found.")
 
         if sha is None:
-            resp: dict = await self.request(self.REPO + "/git/refs/heads/" + self.BRANCH)
+            resp = await self.request(self.REPO + "/git/refs/heads/" + self.BRANCH)
             sha = resp["object"]["sha"]
 
         payload = {"base": self.BRANCH, "head": sha, "commit_message": "Updating bot"}
 
         merge_url = self.MERGE_URL.format(username=self.username)
 
-        resp = await self.request(merge_url, method="POST", payload=payload)
-        if isinstance(resp, dict):
-            return resp
+        resp = await self.request(
+            merge_url,
+            method="POST",
+            payload=payload,
+            return_response=True,
+            read_before_return=True,
+        )
+
+        repo_url = self.BASE + f"/repos/{self.username}/modmail"
+        status_map = {
+            201: "Successful response.",
+            204: "Already merged.",
+            403: "Forbidden.",
+            404: f"Repository '{repo_url}' not found.",
+            409: "There is a merge conflict.",
+            422: "Validation failed.",
+        }
+        # source https://docs.github.com/en/rest/branches/branches#merge-a-branch
+
+        status = resp.status
+        data = await self._get_response_data(resp)
+        if status in (201, 204):
+            return data
+
+        args = (resp.request_info, resp.history)
+        try:
+            # try to get the response error message if any
+            message = data.get("message")
+        except AttributeError:
+            message = None
+        kwargs = {
+            "status": status,
+            "message": message if message else status_map.get(status),
+        }
+        # just raise
+        raise ClientResponseError(*args, **kwargs)
 
     async def fork_repository(self) -> None:
         """
         Forks Modmail's repository.
         """
-        await self.request(self.FORK_URL, method="POST")
+        await self.request(self.FORK_URL, method="POST", return_response=True)
 
     async def has_starred(self) -> bool:
         """
         Checks if shared Modmail.
+
         Returns
         -------
         bool
@@ -187,23 +246,30 @@ class GitHub:
         """
         Stars Modmail's repository.
         """
-        await self.request(self.STAR_URL, method="PUT", headers={"Content-Length": "0"})
+        await self.request(
+            self.STAR_URL,
+            method="PUT",
+            headers={"Content-Length": "0"},
+            return_response=True,
+        )
 
     @classmethod
     async def login(cls, bot) -> "GitHub":
         """
         Logs in to GitHub with configuration variable information.
+
         Parameters
         ----------
         bot : Bot
             The Modmail bot.
+
         Returns
         -------
         GitHub
             The newly created `GitHub` object.
         """
         self = cls(bot, bot.config.get("github_token"))
-        resp: dict = await self.request("https://api.github.com/user")
+        resp: Dict[str, Any] = await self.request(self.BASE + "/user")
         if resp.get("login"):
             self.username = resp["login"]
             self.avatar_url = resp["avatar_url"]
@@ -288,6 +354,9 @@ class ApiClient:
         return NotImplemented
 
     async def get_user_logs(self, user_id: Union[str, int]) -> list:
+        return NotImplemented
+
+    async def find_log_entry(self, key: str) -> list:
         return NotImplemented
 
     async def get_latest_user_logs(self, user_id: Union[str, int]):
@@ -463,6 +532,13 @@ class MongoDBClient(ApiClient):
 
         return await self.logs.find(query, projection).to_list(None)
 
+    async def find_log_entry(self, key: str) -> list:
+        query = {"key": key}
+        projection = {"messages": {"$slice": 5}}
+        logger.debug(f"Retrieving log ID {key}.")
+
+        return await self.logs.find(query, projection).to_list(None)
+
     async def get_latest_user_logs(self, user_id: Union[str, int]):
         query = {"recipient.id": str(user_id), "guild_id": str(self.bot.guild_id), "open": False}
         projection = {"messages": {"$slice": 5}}
@@ -507,7 +583,7 @@ class MongoDBClient(ApiClient):
                 "_id": key,
                 "key": key,
                 "open": True,
-                "created_at": str(datetime.utcnow()),
+                "created_at": str(discord.utils.utcnow()),
                 "closed_at": None,
                 "channel_id": str(channel.id),
                 "guild_id": str(self.bot.guild_id),
@@ -516,14 +592,14 @@ class MongoDBClient(ApiClient):
                     "id": str(recipient.id),
                     "name": recipient.name,
                     "discriminator": recipient.discriminator,
-                    "avatar_url": str(recipient.avatar_url),
+                    "avatar_url": recipient.display_avatar.url,
                     "mod": False,
                 },
                 "creator": {
                     "id": str(creator.id),
                     "name": creator.name,
                     "discriminator": creator.discriminator,
-                    "avatar_url": str(creator.avatar_url),
+                    "avatar_url": creator.display_avatar.url,
                     "mod": isinstance(creator, Member),
                 },
                 "closer": None,
@@ -585,7 +661,7 @@ class MongoDBClient(ApiClient):
                 "id": str(message.author.id),
                 "name": message.author.name,
                 "discriminator": message.author.discriminator,
-                "avatar_url": str(message.author.avatar_url),
+                "avatar_url": message.author.display_avatar.url,
                 "mod": not isinstance(message.channel, DMChannel),
             },
             "content": message.content,
@@ -635,7 +711,7 @@ class MongoDBClient(ApiClient):
                     "id": str(message.author.id),
                     "name": message.author.name,
                     "discriminator": message.author.discriminator,
-                    "avatar_url": str(message.author.avatar_url),
+                    "avatar_url": message.author.display_avatar.url,
                 },
                 "message": message.content,
                 "message_id": str(message_id),
