@@ -13,7 +13,7 @@ from discord.ext.commands import BadArgument
 
 from core._color_data import ALL_COLORS
 from core.models import DMDisabled, InvalidConfigError, Default, getLogger
-from core.time import UserFriendlyTimeSync
+from core.time import UserFriendlyTime
 from core.utils import strtobool
 
 logger = getLogger(__name__)
@@ -21,7 +21,6 @@ load_dotenv()
 
 
 class ConfigManager:
-
     public_keys = {
         # activity
         "twitch_url": "https://www.twitch.tv/discordmodmail/",
@@ -37,6 +36,7 @@ class ConfigManager:
         "account_age": isodate.Duration(),
         "guild_age": isodate.Duration(),
         "thread_cooldown": isodate.Duration(),
+        "log_expiration": isodate.Duration(),
         "reply_without_command": False,
         "anon_reply_without_command": False,
         "plain_reply_without_command": False,
@@ -52,6 +52,8 @@ class ConfigManager:
         "close_emoji": "\N{LOCK}",
         "use_user_id_channel_name": False,
         "use_timestamp_channel_name": False,
+        "use_nickname_channel_name": False,
+        "use_random_channel_name": False,
         "recipient_thread_close": False,
         "thread_show_roles": True,
         "thread_show_account_age": True,
@@ -77,7 +79,7 @@ class ConfigManager:
         "thread_move_notify_mods": False,
         "thread_move_response": "This thread has been moved.",
         "cooldown_thread_title": "Message not sent!",
-        "cooldown_thread_response": "You must wait for {delta} before you can contact me again.",
+        "cooldown_thread_response": "Your cooldown ends {delta}. Try contacting me then.",
         "disabled_new_thread_title": "Not Delivered",
         "disabled_new_thread_response": "We are not accepting new threads.",
         "disabled_new_thread_footer": "Please try again later...",
@@ -91,6 +93,9 @@ class ConfigManager:
         "silent_alert_on_mention": False,
         "show_timestamp": True,
         "anonymous_snippets": False,
+        "plain_snippets": False,
+        "require_close_reason": False,
+        "show_log_url_button": False,
         # group conversations
         "private_added_to_group_title": "New Thread (Group)",
         "private_added_to_group_response": "{moderator.name} has added you to a Modmail thread.",
@@ -118,11 +123,12 @@ class ConfigManager:
         # confirm thread creation
         "confirm_thread_creation": False,
         "confirm_thread_creation_title": "Confirm thread creation",
-        "confirm_thread_response": "React to confirm thread creation which will directly contact the moderators",
+        "confirm_thread_response": "Click the button to confirm thread creation which will directly contact the moderators.",
         "confirm_thread_creation_accept": "\N{WHITE HEAVY CHECK MARK}",
         "confirm_thread_creation_deny": "\N{NO ENTRY SIGN}",
         # regex
         "use_regex_autotrigger": False,
+        "use_hoisted_top_role": True,
     }
 
     private_keys = {
@@ -160,6 +166,8 @@ class ConfigManager:
         "database_type": "mongodb",
         "connection_uri": None,  # replace mongo uri in the future
         "owners": None,
+        "enable_presence_intent": False,
+        "registry_plugins_only": False,
         # bot
         "token": None,
         "enable_plugins": True,
@@ -170,22 +178,28 @@ class ConfigManager:
         "disable_updates": False,
         # Logging
         "log_level": "INFO",
+        "stream_log_format": "plain",
+        "file_log_format": "plain",
+        "discord_log_level": "INFO",
         # data collection
         "data_collection": True,
     }
 
     colors = {"mod_color", "recipient_color", "main_color", "error_color"}
 
-    time_deltas = {"account_age", "guild_age", "thread_auto_close", "thread_cooldown"}
+    time_deltas = {"account_age", "guild_age", "thread_auto_close", "thread_cooldown", "log_expiration"}
 
     booleans = {
         "use_user_id_channel_name",
         "use_timestamp_channel_name",
+        "use_nickname_channel_name",
+        "use_random_channel_name",
         "user_typing",
         "mod_typing",
         "reply_without_command",
         "anon_reply_without_command",
         "plain_reply_without_command",
+        "show_log_url_button",
         "recipient_thread_close",
         "thread_auto_close_silently",
         "thread_move_notify",
@@ -205,10 +219,15 @@ class ConfigManager:
         "update_notifications",
         "thread_contact_silently",
         "anonymous_snippets",
+        "plain_snippets",
+        "require_close_reason",
         "recipient_thread_close",
         "thread_show_roles",
         "thread_show_account_age",
         "thread_show_join_age",
+        "use_hoisted_top_role",
+        "enable_presence_intent",
+        "registry_plugins_only",
     }
 
     enums = {
@@ -285,7 +304,7 @@ class ConfigManager:
     def __delitem__(self, key: str) -> None:
         return self.remove(key)
 
-    def get(self, key: str, convert=True) -> typing.Any:
+    def get(self, key: str, *, convert: bool = True) -> typing.Any:
         key = key.lower()
         if key not in self.all_keys:
             raise InvalidConfigError(f'Configuration "{key}" is invalid.')
@@ -354,7 +373,7 @@ class ConfigManager:
 
         return value
 
-    def set(self, key: str, item: typing.Any, convert=True) -> None:
+    async def set(self, key: str, item: typing.Any, convert=True) -> None:
         if not convert:
             return self.__setitem__(key, item)
 
@@ -388,8 +407,8 @@ class ConfigManager:
                 isodate.parse_duration(item)
             except isodate.ISO8601Error:
                 try:
-                    converter = UserFriendlyTimeSync()
-                    time = converter.convert(None, item)
+                    converter = UserFriendlyTime()
+                    time = await converter.convert(None, item, now=discord.utils.utcnow())
                     if time.arg:
                         raise ValueError
                 except BadArgument as exc:
@@ -400,7 +419,8 @@ class ConfigManager:
                         "Unrecognized time, please use ISO-8601 duration format "
                         'string or a simpler "human readable" time.'
                     )
-                item = isodate.duration_isoformat(time.dt - converter.now)
+                now = discord.utils.utcnow()
+                item = isodate.duration_isoformat(time.dt - now)
             return self.__setitem__(key, item)
 
         if key in self.booleans:
