@@ -1,8 +1,8 @@
 import typing
-import asyncio
 
-from discord import User, Reaction, Message, Embed
-from discord import HTTPException, InvalidArgument
+import discord
+from discord import Message, Embed, ButtonStyle, Interaction
+from discord.ui import View, Button, Select
 from discord.ext import commands
 
 
@@ -33,8 +33,12 @@ class PaginatorSession:
         The `Message` of the `Embed`.
     current : int
         The current page number.
-    reaction_map : Dict[str, method]
-        A mapping for reaction to method.
+    callback_map : Dict[str, method]
+        A mapping for text to method.
+    view : PaginatorView
+        The view that is sent along with the base message.
+    select_menu : Select
+        A select menu that will be added to the View.
     """
 
     def __init__(self, ctx: commands.Context, *pages, **options):
@@ -45,40 +49,18 @@ class PaginatorSession:
         self.current = 0
         self.pages = list(pages)
         self.destination = options.get("destination", ctx)
-        self.reaction_map = {
-            "⏮": self.first_page,
-            "◀": self.previous_page,
-            "▶": self.next_page,
-            "⏭": self.last_page,
-            "🛑": self.close,
+        self.view = None
+        self.select_menu = None
+
+        self.callback_map = {
+            "<<": self.first_page,
+            "<": self.previous_page,
+            ">": self.next_page,
+            ">>": self.last_page,
         }
+        self._buttons_map = {"<<": None, "<": None, ">": None, ">>": None}
 
-    def add_page(self, item) -> None:
-        """
-        Add a page.
-        """
-        raise NotImplementedError
-
-    async def create_base(self, item) -> None:
-        """
-        Create a base `Message`.
-        """
-        await self._create_base(item)
-
-        if len(self.pages) == 1:
-            self.running = False
-            return
-
-        self.running = True
-        for reaction in self.reaction_map:
-            if len(self.pages) == 2 and reaction in "⏮⏭":
-                continue
-            await self.ctx.bot.add_reaction(self.base, reaction)
-
-    async def _create_base(self, item) -> None:
-        raise NotImplementedError
-
-    async def show_page(self, index: int) -> None:
+    async def show_page(self, index: int) -> typing.Optional[typing.Dict]:
         """
         Show a page by page number.
 
@@ -92,74 +74,96 @@ class PaginatorSession:
 
         self.current = index
         page = self.pages[index]
+        result = None
 
         if self.running:
-            await self._show_page(page)
+            result = self._show_page(page)
         else:
             await self.create_base(page)
 
-    async def _show_page(self, page):
+        self.update_disabled_status()
+        return result
+
+    def update_disabled_status(self):
+        if self.current == self.first_page():
+            # disable << button
+            if self._buttons_map["<<"] is not None:
+                self._buttons_map["<<"].disabled = True
+
+            if self._buttons_map["<"] is not None:
+                self._buttons_map["<"].disabled = True
+        else:
+            if self._buttons_map["<<"] is not None:
+                self._buttons_map["<<"].disabled = False
+
+            if self._buttons_map["<"] is not None:
+                self._buttons_map["<"].disabled = False
+
+        if self.current == self.last_page():
+            # disable >> button
+            if self._buttons_map[">>"] is not None:
+                self._buttons_map[">>"].disabled = True
+
+            if self._buttons_map[">"] is not None:
+                self._buttons_map[">"].disabled = True
+        else:
+            if self._buttons_map[">>"] is not None:
+                self._buttons_map[">>"].disabled = False
+
+            if self._buttons_map[">"] is not None:
+                self._buttons_map[">"].disabled = False
+
+    async def create_base(self, item) -> None:
+        """
+        Create a base `Message`.
+        """
+        if len(self.pages) == 1:
+            self.view = None
+            self.running = False
+        else:
+            self.view = PaginatorView(self, timeout=self.timeout)
+            self.update_disabled_status()
+            self.running = True
+
+        await self._create_base(item, self.view)
+
+    async def _create_base(self, item, view: View) -> None:
         raise NotImplementedError
 
-    def react_check(self, reaction: Reaction, user: User) -> bool:
-        """
+    def _show_page(self, page):
+        raise NotImplementedError
 
-        Parameters
-        ----------
-        reaction : Reaction
-            The `Reaction` object of the reaction.
-        user : User
-            The `User` or `Member` object of who sent the reaction.
+    def first_page(self):
+        """Returns the index of the first page"""
+        return 0
 
-        Returns
-        -------
-        bool
-        """
-        return (
-            reaction.message.id == self.base.id
-            and user.id == self.ctx.author.id
-            and reaction.emoji in self.reaction_map.keys()
-        )
+    def next_page(self):
+        """Returns the index of the next page"""
+        return min(self.current + 1, self.last_page())
+
+    def previous_page(self):
+        """Returns the index of the previous page"""
+        return max(self.current - 1, self.first_page())
+
+    def last_page(self):
+        """Returns the index of the last page"""
+        return len(self.pages) - 1
 
     async def run(self) -> typing.Optional[Message]:
         """
         Starts the pagination session.
-
-        Returns
-        -------
-        Optional[Message]
-            If it's closed before running ends.
         """
         if not self.running:
             await self.show_page(self.current)
-        while self.running:
-            try:
-                reaction, user = await self.ctx.bot.wait_for(
-                    "reaction_add", check=self.react_check, timeout=self.timeout
-                )
-            except asyncio.TimeoutError:
-                return await self.close(delete=False)
-            else:
-                action = self.reaction_map.get(reaction.emoji)
-                await action()
-            try:
-                await self.base.remove_reaction(reaction, user)
-            except (HTTPException, InvalidArgument):
-                pass
 
-    async def previous_page(self) -> None:
-        """
-        Go to the previous page.
-        """
-        await self.show_page(self.current - 1)
+            if self.view is not None:
+                await self.view.wait()
 
-    async def next_page(self) -> None:
-        """
-        Go to the next page.
-        """
-        await self.show_page(self.current + 1)
+            await self.close(delete=False)
 
-    async def close(self, delete: bool = True) -> typing.Optional[Message]:
+    async def close(
+        self, delete: bool = True, *, interaction: Interaction = None
+    ) -> typing.Optional[Message]:
         """
         Closes the pagination session.
 
@@ -174,30 +178,127 @@ class PaginatorSession:
         Optional[Message]
             If `delete` is `True`.
         """
-        self.running = False
+        if self.running:
+            sent_emoji, _ = await self.ctx.bot.retrieve_emoji()
+            await self.ctx.bot.add_reaction(self.ctx.message, sent_emoji)
 
-        sent_emoji, _ = await self.ctx.bot.retrieve_emoji()
-        await self.ctx.bot.add_reaction(self.ctx.message, sent_emoji)
+            if interaction:
+                message = interaction.message
+            else:
+                message = self.base
 
-        if delete:
-            return await self.base.delete()
+            self.running = False
 
-        try:
-            await self.base.clear_reactions()
-        except HTTPException:
-            pass
+            if self.view is not None:
+                self.view.stop()
+                if delete:
+                    await message.delete()
+                else:
+                    self.view.clear_items()
+                    await message.edit(view=self.view)
 
-    async def first_page(self) -> None:
-        """
-        Go to the first page.
-        """
-        await self.show_page(0)
 
-    async def last_page(self) -> None:
-        """
-        Go to the last page.
-        """
-        await self.show_page(len(self.pages) - 1)
+class PaginatorView(View):
+    """
+    View that is used for pagination.
+
+    Parameters
+    ----------
+    handler : PaginatorSession
+        The paginator session that spawned this view.
+    timeout : float
+        How long to wait for before the session closes.
+
+    Attributes
+    ----------
+    handler : PaginatorSession
+        The paginator session that spawned this view.
+    timeout : float
+        How long to wait for before the session closes.
+    """
+
+    def __init__(self, handler: PaginatorSession, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.handler = handler
+        self.clear_items()  # clear first so we can control the order
+        self.fill_items()
+
+    @discord.ui.button(label="Stop", style=ButtonStyle.danger)
+    async def stop_button(self, interaction: Interaction, button: Button):
+        await self.handler.close(interaction=interaction)
+
+    def fill_items(self):
+        if self.handler.select_menu is not None:
+            self.add_item(self.handler.select_menu)
+
+        for label, callback in self.handler.callback_map.items():
+            if len(self.handler.pages) == 2 and label in ("<<", ">>"):
+                continue
+
+            if label in ("<<", ">>"):
+                style = ButtonStyle.secondary
+            else:
+                style = ButtonStyle.primary
+
+            button = PageButton(self.handler, callback, label=label, style=style)
+
+            self.handler._buttons_map[label] = button
+            self.add_item(button)
+        self.add_item(self.stop_button)
+
+    async def interaction_check(self, interaction: Interaction):
+        """Only allow the message author to interact"""
+        if interaction.user != self.handler.ctx.author:
+            await interaction.response.send_message(
+                "Only the original author can control this!", ephemeral=True
+            )
+            return False
+        return True
+
+
+class PageButton(Button):
+    """
+    A button that has a callback to jump to the next page
+
+    Parameters
+    ----------
+    handler : PaginatorSession
+        The paginator session that spawned this view.
+    page_callback : Callable
+        A callable that returns an int of the page to go to.
+
+    Attributes
+    ----------
+    handler : PaginatorSession
+        The paginator session that spawned this view.
+    page_callback : Callable
+        A callable that returns an int of the page to go to.
+    """
+
+    def __init__(self, handler, page_callback, **kwargs):
+        super().__init__(**kwargs)
+        self.handler = handler
+        self.page_callback = page_callback
+
+    async def callback(self, interaction: Interaction):
+        kwargs = await self.handler.show_page(self.page_callback())
+        await interaction.response.edit_message(**kwargs, view=self.view)
+
+
+class PageSelect(Select):
+    def __init__(self, handler: PaginatorSession, pages: typing.List[typing.Tuple[str]]):
+        self.handler = handler
+        options = []
+        for n, (label, description) in enumerate(pages):
+            options.append(discord.SelectOption(label=label, description=description, value=str(n)))
+
+        options = options[:25]  # max 25 options
+        super().__init__(placeholder="Select a page", min_values=1, max_values=1, options=options)
+
+    async def callback(self, interaction: Interaction):
+        page = int(self.values[0])
+        kwargs = await self.handler.show_page(page)
+        await interaction.response.edit_message(**kwargs, view=self.view)
 
 
 class EmbedPaginatorSession(PaginatorSession):
@@ -205,11 +306,42 @@ class EmbedPaginatorSession(PaginatorSession):
         super().__init__(ctx, *embeds, **options)
 
         if len(self.pages) > 1:
+            select_options = []
+            create_select = True
             for i, embed in enumerate(self.pages):
                 footer_text = f"Page {i + 1} of {len(self.pages)}"
                 if embed.footer.text:
                     footer_text = footer_text + " • " + embed.footer.text
-                embed.set_footer(text=footer_text, icon_url=embed.footer.icon_url)
+
+                if embed.footer.icon:
+                    icon_url = embed.footer.icon.url
+                else:
+                    icon_url = None
+                embed.set_footer(text=footer_text, icon_url=icon_url)
+
+                # select menu
+                if embed.author.name:
+                    title = embed.author.name[:30].strip()
+                    if len(embed.author.name) > 30:
+                        title += "..."
+                else:
+                    title = embed.title[:30].strip()
+                    if len(embed.title) > 30:
+                        title += "..."
+                    if not title:
+                        create_select = False
+
+                if embed.description:
+                    description = embed.description[:40].replace("*", "").replace("`", "").strip()
+                    if len(embed.description) > 40:
+                        description += "..."
+                else:
+                    description = ""
+                select_options.append((title, description))
+
+            if create_select:
+                if len(set(x[0] for x in select_options)) != 1:  # must have unique authors
+                    self.select_menu = PageSelect(self, select_options)
 
     def add_page(self, item: Embed) -> None:
         if isinstance(item, Embed):
@@ -217,11 +349,11 @@ class EmbedPaginatorSession(PaginatorSession):
         else:
             raise TypeError("Page must be an Embed object.")
 
-    async def _create_base(self, item: Embed) -> None:
-        self.base = await self.destination.send(embed=item)
+    async def _create_base(self, item: Embed, view: View) -> None:
+        self.base = await self.destination.send(embed=item, view=view)
 
-    async def _show_page(self, page):
-        await self.base.edit(embed=page)
+    def _show_page(self, page):
+        return dict(embed=page)
 
 
 class MessagePaginatorSession(PaginatorSession):
@@ -241,12 +373,18 @@ class MessagePaginatorSession(PaginatorSession):
             footer_text = f"Page {self.current+1} of {len(self.pages)}"
             if self.footer_text:
                 footer_text = footer_text + " • " + self.footer_text
-            self.embed.set_footer(text=footer_text, icon_url=self.embed.footer.icon_url)
 
-    async def _create_base(self, item: str) -> None:
-        self._set_footer()
-        self.base = await self.ctx.send(content=item, embed=self.embed)
+            if self.embed.footer.icon:
+                icon_url = self.embed.footer.icon.url
+            else:
+                icon_url = None
 
-    async def _show_page(self, page) -> None:
+            self.embed.set_footer(text=footer_text, icon_url=icon_url)
+
+    async def _create_base(self, item: str, view: View) -> None:
         self._set_footer()
-        await self.base.edit(content=page, embed=self.embed)
+        self.base = await self.ctx.send(content=item, embed=self.embed, view=view)
+
+    def _show_page(self, page) -> typing.Dict:
+        self._set_footer()
+        return dict(content=page, embed=self.embed)
